Module gfx::core

Modules

draw
Command Buffer device interface

dummy
Dummy backend implementation to test the code for compile errors outside of the graphics development environment.

factory
Resource factory.

format
Universal format specification. Applicable to textures, views, and vertex buffers.

handle
Device resource handles

mapping
Memory mapping

pso
Pipeline State Objects

shade
Shader handling.

state
Fixed-function hardware state.

target
Render target specification.

tex
Texture creation and modification.

Structs

Capabilities
Features that the device supports.

DomainShader

GeometryShader

HullShader

PixelShader

VertexShader

Enums

IndexType
A type of each index value in the mesh's index buffer

Primitive
Describes what geometric primitives are created from vertex data.

ShaderSet
A complete set of shaders to link a program.

Constants

MAX_COLOR_TARGETS
Compile-time maximum number of color targets.

MAX_CONSTANT_BUFFERS
Compile-time maximum number of constant buffers.

MAX_RESOURCE_VIEWS
Compile-time maximum number of shader resource views (SRV).

MAX_SAMPLERS
Compile-time maximum number of samplers.

MAX_UNORDERED_VIEWS
Compile-time maximum number of unordered access views (UAV).

MAX_VERTEX_ATTRIBUTES
Compile-time maximum number of vertex attributes.

Traits

Device
An interface for performing draw calls using a specific graphics API

DeviceFence
Extension to the Device that allows for submitting of commands around a fence

Factory

Resources
Resources pertaining to a specific API.

Type Definitions

AttributeSlot

ColorSlot

ConstantBufferSlot

InstanceCount

ResourceViewSlot

SamplerSlot

UnorderedViewSlot

VertexCount