Module gfx::core::pso

Pipeline State Objects

Structs

ColorInfo
Color output configuration of the PSO.

ConstantBufferParam
A constant buffer run-time parameter for PSO.

CreationError
Error types happening upon PSO creation on the device side.

DepthStencilInfo
Depth and stencil state of the PSO.

Descriptor
All the information surrounding a shader program that is required for PSO creation, including the formats of vertex buffers and pixel targets;

Element
A struct element descriptor.

PixelTargetSet
A complete set of render targets to be used for pixel export in PSO.

ResourceViewParam
A shader resource view (SRV) run-time parameter for PSO.

SamplerParam
A sampler run-time parameter for PSO.

UnorderedViewParam
An unordered access view (UAV) run-time parameter for PSO.

VertexBufferSet
A complete set of vertex buffers to be used for vertex import in PSO.

Type Definitions

AttributeDesc

BufferOffset

ColorTargetDesc

DepthStencilDesc

ElemOffset

ElemStride

InstanceRate