Trait gfx::core::Device

pub trait Device {
    type Resources: Resources;
    type CommandBuffer: CommandBuffer<Self::Resources>;
    fn get_capabilities(&self) -> &Capabilities;
    fn pin_submitted_resources(&mut self, &Manager<Self::Resources>);
    fn submit(&mut self, &mut Self::CommandBuffer);
    fn cleanup(&mut self);
}

An interface for performing draw calls using a specific graphics API

Associated Types

type Resources: Resources

Associated resources type.

type CommandBuffer: CommandBuffer<Self::Resources>

Associated command buffer type.

Required Methods

fn get_capabilities(&self) -> &Capabilities

Returns the capabilities available to the specific API implementation.

fn pin_submitted_resources(&mut self, &Manager<Self::Resources>)

Pin everything from this handle manager to live for a frame.

fn submit(&mut self, &mut Self::CommandBuffer)

Submit a command buffer for execution.

fn cleanup(&mut self)

Cleanup unused resources, to be called between frames.