Module gfx::device

Graphics device. Not meant for direct use.

Modules

attrib
Vertex attribute types.

draw
Command Buffer device interface

handle
Device resource handles

mapping
Memory mapping

shade
Shader handling.

state
Fixed-function hardware state.

target
Render target specification.

tex
Texture creation and modification.

Structs

BufferInfo
An information block that is immutable and associated with each buffer

Capabilities
Features that the device supports.

NotSupported
Generic error for features that are not supported by the device capabilities.

Enums

BufferRole
Role of the memory buffer. GLES doesn't chaning bind points for buffers.

BufferUpdateError
An error happening on buffer updates.

BufferUsage
A hint as to how this buffer will be used.

MapAccess
Specifies the access allowed to a buffer mapping.

PrimitiveType
Describes what geometric primitives are created from vertex data.

Traits

Device
An interface for performing draw calls using a specific graphics API

Factory

Resources
Resources pertaining to a specific API.

Functions

as_byte_slice
Treat a given slice as &[u8] for the given function call

Type Definitions

AttributeSlot
Slot for an attribute.

IndexType
A type of each index value in the mesh's index buffer

InstanceCount
Draw number of instances

SubmitInfo
All the data needed simultaneously for submitting a command buffer for execution on a device.

TextureSlot
Slot a texture can be bound to.

UniformBlockIndex
Index of a uniform block.

UniformBufferSlot
Slot for a uniform buffer object.

VertexCount
Draw vertex count.