Module gfx::device::attrib

Vertex attribute types.

Modules

format
A service module for deriving ToFormat for primitive types.

Structs

FixedPoint
Fixed-point version of integer attributes.

Floater
Floating-point version of integer attributes.

Format
Complete format of a vertex attribute.

Enums

FloatSize
The size of a floating point attribute, in bits.

FloatSubType
Type of a floating point attribute on the shader side.

IntSize
The size of an integer attribute, in bits.

IntSubType
Describes how an integer value is interpreted by the shader.

SignFlag
The signedness of an attribute.

Type
The type of an attribute.

Type Definitions

Count
Number of elements per attribute, only 1 to 4 are supported

InstanceRate
The number of instances between each subsequent attribute value

Offset
Offset of an attribute from the start of the buffer, in bytes

Stride
Offset between attribute values, in bytes