Module gfx::device::draw

Command Buffer device interface

Structs

DataBuffer
A buffer of data accompanying the commands. It can be vertex data, texture updates, uniform blocks, or even some draw states.

DataPointer
The place of some data in the data buffer.

Enums

Access
Type of the frame buffer access.

Gamma
Type of the gamma transformation for framebuffer writes.

Target
When rendering, each "output" of the fragment shader goes to a specific target. A Plane can be bound to a target, causing writes to that target to affect the Plane.

Traits

CommandBuffer
An interface of the abstract command buffer. It collects commands in an efficient API-specific manner, to be ready for execution on the device.

Type Definitions

InstanceOption
Optional instance parameters