Trait gfx::device::draw::CommandBuffer

pub trait CommandBuffer<R: Resources> {
    fn new() -> Self;
    fn clear(&mut self);
    fn bind_program(&mut self, R::Program);
    fn bind_array_buffer(&mut self, R::ArrayBuffer);
    fn bind_attribute(&mut self, AttributeSlot, R::Buffer, Format);
    fn bind_index(&mut self, R::Buffer);
    fn bind_frame_buffer(&mut self, Access, R::FrameBuffer, Gamma);
    fn unbind_target(&mut self, Access, Target);
    fn bind_target_surface(&mut self, Access, Target, R::Surface);
    fn bind_target_texture(&mut self, Access, Target, R::Texture, Level, Option<Layer>);
    fn bind_uniform_block(&mut self, R::Program, UniformBufferSlot, UniformBlockIndex, R::Buffer);
    fn bind_uniform(&mut self, Location, UniformValue);
    fn bind_texture(&mut self, TextureSlot, Kind, R::Texture, Option<(R::Sampler, SamplerInfo)>);
    fn set_draw_color_buffers(&mut self, usize);
    fn set_primitive(&mut self, Primitive);
    fn set_viewport(&mut self, Rect);
    fn set_multi_sample(&mut self, Option<MultiSample>);
    fn set_scissor(&mut self, Option<Rect>);
    fn set_depth_stencil(&mut self, Option<Depth>, Option<Stencil>, CullFace);
    fn set_blend(&mut self, Option<Blend>);
    fn set_color_mask(&mut self, ColorMask);
    fn update_buffer(&mut self, R::Buffer, DataPointer, usize);
    fn update_texture(&mut self, Kind, R::Texture, ImageInfo, DataPointer);
    fn call_clear(&mut self, ClearData, Mask);
    fn call_draw(&mut self, PrimitiveType, VertexCount, VertexCount, InstanceOption);
    fn call_draw_indexed(&mut self, PrimitiveType, IndexType, VertexCount, VertexCount, VertexCount, InstanceOption);
    fn call_blit(&mut self, Rect, Rect, Mirror, Mask);
}

An interface of the abstract command buffer. It collects commands in an efficient API-specific manner, to be ready for execution on the device.

Required Methods

fn new() -> Self

An empty constructor

fn clear(&mut self)

Clear the command buffer contents, retain the allocated storage

fn bind_program(&mut self, R::Program)

Bind a shader program

fn bind_array_buffer(&mut self, R::ArrayBuffer)

Bind an array buffer object

fn bind_attribute(&mut self, AttributeSlot, R::Buffer, Format)

Bind a vertex attribute

fn bind_index(&mut self, R::Buffer)

Bind an index buffer

fn bind_frame_buffer(&mut self, Access, R::FrameBuffer, Gamma)

Bind a frame buffer object

fn unbind_target(&mut self, Access, Target)

Unbind any surface from the specified target slot

fn bind_target_surface(&mut self, Access, Target, R::Surface)

Bind a surface to the specified target slot

fn bind_target_texture(&mut self, Access, Target, R::Texture, Level, Option<Layer>)

Bind a level of the texture to the specified target slot

fn bind_uniform_block(&mut self, R::Program, UniformBufferSlot, UniformBlockIndex, R::Buffer)

Bind a uniform block

fn bind_uniform(&mut self, Location, UniformValue)

Bind a single uniform in the default block

fn bind_texture(&mut self, TextureSlot, Kind, R::Texture, Option<(R::Sampler, SamplerInfo)>)

Bind a texture

fn set_draw_color_buffers(&mut self, usize)

Select, which color buffers are going to be targetted by the shader

fn set_primitive(&mut self, Primitive)

Set primitive topology

fn set_viewport(&mut self, Rect)

Set viewport rectangle

fn set_multi_sample(&mut self, Option<MultiSample>)

Set multi-sampling state

fn set_scissor(&mut self, Option<Rect>)

Set scissor test

fn set_depth_stencil(&mut self, Option<Depth>, Option<Stencil>, CullFace)

Set depth and stencil states

fn set_blend(&mut self, Option<Blend>)

Set blend state

fn set_color_mask(&mut self, ColorMask)

Set output color mask for all targets

fn update_buffer(&mut self, R::Buffer, DataPointer, usize)

Update a vertex/index/uniform buffer

fn update_texture(&mut self, Kind, R::Texture, ImageInfo, DataPointer)

Update a texture region

fn call_clear(&mut self, ClearData, Mask)

Clear target surfaces

fn call_draw(&mut self, PrimitiveType, VertexCount, VertexCount, InstanceOption)

Draw a primitive

fn call_draw_indexed(&mut self, PrimitiveType, IndexType, VertexCount, VertexCount, VertexCount, InstanceOption)

Draw a primitive with index buffer

fn call_blit(&mut self, Rect, Rect, Mirror, Mask)

Blit from one target to another