Enum gfx::device::BufferUsage

pub enum BufferUsage {
    Static,
    Dynamic,
    Stream,
}

A hint as to how this buffer will be used.

The nature of these hints make them very implementation specific. Different drivers on different hardware will handle them differently. Only careful profiling will tell which is the best to use for a specific buffer.

Variants

Static

Once uploaded, this buffer will rarely change, but will be read from often.

Dynamic

This buffer will be updated "frequently", and will be read from multiple times between updates.

Stream

This buffer always or almost always be updated after each read.