Module gfx::device::handle

Device resource handles

Structs

ArrayBuffer
Array Buffer Handle

Buffer
Type-safe buffer handle

FrameBuffer
Frame Buffer Handle

Manager
Stores reference-counted resources used in a command buffer. Seals actual resource names behind the interface, automatically referencing them both by the Factory on resource creation and the Renderer during CommandBuffer population.

Program
Program Handle

RawBuffer
Raw (untyped) Buffer Handle

Sampler
Sampler Handle

Shader
Shader Handle

Surface
Surface Handle

Texture
Texture Handle

Traits

Producer
A service trait to be used by the device implementation