Struct gfx::device::handle::Manager

pub struct Manager<R: Resources> {
    // some fields omitted
}

Stores reference-counted resources used in a command buffer. Seals actual resource names behind the interface, automatically referencing them both by the Factory on resource creation and the Renderer during CommandBuffer population.

Methods

impl<R: Resources> Manager<R>

fn new() -> Manager<R>

Create a new handle manager

fn clear(&mut self)

Clear all references

fn extend(&mut self, other: &Manager<R>)

Extend with all references of another handle manager

fn count(&self) -> usize

Count the total number of referenced resources

fn ref_buffer(&mut self, handle: &RawBuffer<R>) -> R::Buffer

Reference a buffer

fn ref_array_buffer(&mut self, handle: &ArrayBuffer<R>) -> R::ArrayBuffer

Reference am array buffer

fn ref_shader(&mut self, handle: &Shader<R>) -> R::Shader

Reference a shader

fn ref_program(&mut self, handle: &Program<R>) -> R::Program

Reference a program

fn ref_frame_buffer(&mut self, handle: &FrameBuffer<R>) -> R::FrameBuffer

Reference a frame buffer

fn ref_surface(&mut self, handle: &Surface<R>) -> R::Surface

Reference a surface

fn ref_texture(&mut self, handle: &Texture<R>) -> R::Texture

Reference a texture

fn ref_sampler(&mut self, handle: &Sampler<R>) -> R::Sampler

Reference a sampler