Module gfx::device::mapping

Memory mapping

Structs

RW
A handle to a complete readable/writable map, which can be sliced both ways.

Readable
A handle to a readable map, which can be sliced.

Writable
A handle to a writable map, which only allows setting elements.

Traits

Builder
A service trait with methods for mapping already implemented. To be used by device back ends.

Raw
Unsafe operations for a buffer mapping