Module gfx::device::shade

Shader handling.

Structs

Attribute
Vertex information that a shader takes as input.

BlockVar
A uniform block.

ProgramInfo
Metadata about a program.

SamplerVar
Sampler, a type of shader parameter representing a texture that can be sampled.

UniformVar
Uniform, a type of shader parameter representing data passed to the program.

Enums

BaseType
Base type of this shader parameter.

CompatibilityError
Error type for trying to store a UniformValue in a UniformVar.

ContainerType
Number of components this parameter represents.

CreateProgramError
An error type for creating programs.

CreateShaderError
An error type for creating shaders.

IsArray
Whether the sampler samples an array texture.

IsMultiSample
Whether the sampler samples a multisample texture.

IsRect
Whether the sampler samples a rectangle texture.

IsShadow
Whether the sampler samples a shadow texture (texture with a depth comparison)

MatrixFormat
Whether the matrix is column or row major.

SamplerType
What texture type this sampler samples from.

ShaderModel
Shader model supported by the device, corresponds to the HLSL shader models.

Stage
Which program stage this shader represents.

UniformValue
A value that can be uploaded to the device as a uniform.

Type Definitions

Dimension
Number of components in a container type (vectors/matrices)

Location
Location of a parameter in the program.