Struct gfx::device::shade::UniformVar

pub struct UniformVar {
    pub name: String,
    pub location: Location,
    pub count: usize,
    pub base_type: BaseType,
    pub container: ContainerType,
}

Uniform, a type of shader parameter representing data passed to the program.

Fields

name

Name of this uniform.

location

Location of this uniform in the program.

count

Number of elements this uniform represents.

base_type

Type that this uniform is composed of

container

"Scalarness" of this uniform.

Methods

impl UniformVar

fn is_compatible(&self, value: &UniformValue) -> Result<(), CompatibilityError>

Whether a value is compatible with this variable. That is, whether the value can be stored in this variable.