Module gfx::device::state

Fixed-function hardware state.

Configures the primitive assembly (PA), rasterizer, and output merger (OM) blocks.

Structs

Blend

BlendChannel

ColorMask

Depth
Depth test state.

MultiSample
Multi-sampling rasterization mode

Offset
How to offset vertices in screen space, if at all.

Primitive
Primitive rasterization state. Note that GL allows different raster method to be used for front and back, while this abstraction does not.

Stencil
Complete stencil state, specifying how to handle the front and back side of a face.

StencilSide
Complete stencil state for a given side of a face.

Enums

BlendValue

Comparison
A pixel-wise comparison function.

CullFace
Which face, if any, to cull.

Equation

Factor

FrontFace
The front face winding order of a set of vertices.

RasterMethod
How to rasterize a primitive.

StencilOp
Stencil mask operation.

Constants

ALPHA

BLUE

GREEN

MASK_ALL

MASK_NONE

RED

Type Definitions

LineWidth

OffsetFactor

OffsetUnits