Enum gfx::device::tex::FilterMethod

pub enum FilterMethod {
    Scale,
    Mipmap,
    Bilinear,
    Trilinear,
    Anisotropic(u8),
}

How to filter the texture when sampling. They correspond to increasing levels of quality, but also cost. They "layer" on top of each other: it is not possible to have bilinear filtering without mipmapping, for example.

These names are somewhat poor, in that "bilinear" is really just doing linear filtering on each axis, and it is only bilinear in the case of 2D textures. Similarly for trilinear, it is really Quadralinear(?) for 3D textures. Alas, these names are simple, and match certain intuitions ingrained by many years of public use of inaccurate terminology.

Variants

Scale

The dumbest filtering possible, nearest-neighbor interpolation.

Mipmap

Add simple mipmapping.

Bilinear

Sample multiple texels within a single mipmap level to increase quality.

Trilinear

Sample multiple texels across two mipmap levels to increase quality.

Anisotropic

Anisotropic filtering with a given "max", must be between 1 and 16, inclusive.