Enum gfx::device::tex::Kind

pub enum Kind {
    D1,
    D1Array,
    D2,
    D2Array,
    D2MultiSample(AaMode),
    D2MultiSampleArray(AaMode),
    Cube(CubeFace),
    D3,
}

Specifies how a given texture may be used. The available texture types are restricted by what Metal exposes, though this could conceivably be extended in the future. Note that a single texture can only ever be of one kind. A texture created as Texture2D will forever be Texture2D.

Variants

D1

A single row of texels.

D1Array

An array of rows of texels. Equivalent to Texture2D except that texels in a different row are not sampled.

D2

A traditional 2D texture, with rows arranged contiguously.

D2Array

An array of 2D textures. Equivalent to Texture3D except that texels in a different depth level are not sampled.

D2MultiSample

A multi-sampled 2D texture. Each pixel may have more than one data value (sample) associated with it.

D2MultiSampleArray

A array of multi-sampled 2D textures.

Cube

A set of 6 2D textures, one for each face of a cube.

When creating a cube texture, the face is ignored, and storage for all 6 faces is created. When updating, only the face specified is updated.

D3

A volume texture, with each 2D layer arranged contiguously.

Methods

impl Kind

fn get_aa_mode(&self) -> Option<AaMode>

Return the anti-aliasing mode of the texture