Trait gfx::device::DeviceFence

pub trait DeviceFence<R: Resources>: Device<Resources=R> where Self::CommandBuffer: CommandBuffer<R> {
    fn fenced_submit(&mut self, SubmitInfo<Self>, after: Option<Fence<R>>) -> Fence<R>;
    fn fence_wait(&mut self, fence: &Fence<R>);
}

Extension to the Device that allows for submitting of commands around a fence

Required Methods

fn fenced_submit(&mut self, SubmitInfo<Self>, after: Option<Fence<R>>) -> Fence<R>

Submit a command buffer to the stream creating a fence the fence is signaled after the GPU has executed all commands in the buffer

fn fence_wait(&mut self, fence: &Fence<R>)

Wait on the supplied fence stalling the current thread until the fence is satisfied