Struct gfx::display_list::BaseDisplayItem

pub struct BaseDisplayItem {
    pub bounds: Rect<Au>,
    pub metadata: DisplayItemMetadata,
    pub clip: ClippingRegion,
    pub section: DisplayListSection,
    pub stacking_context_id: StackingContextId,
}

Information common to all display items.

Fields
bounds: Rect<Au>

The boundaries of the display item, in layer coordinates.

metadata: DisplayItemMetadata

Metadata attached to this display item.

clip: ClippingRegion

The region to clip to.

section: DisplayListSection

The section of the display list that this item belongs to.

stacking_context_id: StackingContextId

The id of the stacking context this item belongs to.

Methods

impl BaseDisplayItem

fn new(bounds: &Rect<Au>, metadata: DisplayItemMetadata, clip: &ClippingRegion, section: DisplayListSection, stacking_context_id: StackingContextId) -> BaseDisplayItem