Struct gfx::display_list::ClippingRegion

pub struct ClippingRegion {
    pub main: Rect<Au>,
    pub complex: Vec<ComplexClippingRegion>,
}

A clipping region for a display item. Currently, this can describe rectangles, rounded rectangles (for border-radius), or arbitrary intersections of the two. Arbitrary transforms are not supported because those are handled by the higher-level StackingContext abstraction.

Fields

main

The main rectangular region. This does not include any corners.

complex

Any complex regions.

TODO(pcwalton): Atomically reference count these? Not sure if it's worth the trouble. Measure and follow up.

Methods

impl ClippingRegion

fn empty() -> ClippingRegion

Returns an empty clipping region that, if set, will result in no pixels being visible.

fn max() -> ClippingRegion

Returns an all-encompassing clipping region that clips no pixels out.

fn from_rect(rect: &Rect<Au>) -> ClippingRegion

Returns a clipping region that represents the given rectangle.

fn intersect_rect(self, rect: &Rect<Au>) -> ClippingRegion

Returns the intersection of this clipping region and the given rectangle.

TODO(pcwalton): This could more eagerly eliminate complex clipping regions, at the cost of complexity.

fn might_be_nonempty(&self) -> bool

Returns true if this clipping region might be nonempty. This can return false positives, but never false negatives.

fn might_intersect_point(&self, point: &Point2D<Au>) -> bool

Returns true if this clipping region might contain the given point and false otherwise. This is a quick, not a precise, test; it can yield false positives.

fn might_intersect_rect(&self, rect: &Rect<Au>) -> bool

Returns true if this clipping region might intersect the given rectangle and false otherwise. This is a quick, not a precise, test; it can yield false positives.

fn bounding_rect(&self) -> Rect<Au>

Returns a bounding rect that surrounds this entire clipping region.

fn intersect_with_rounded_rect(self, rect: &Rect<Au>, radii: &BorderRadii<Au>) -> ClippingRegion

Intersects this clipping region with the given rounded rectangle.

fn translate(&self, delta: &Point2D<Au>) -> ClippingRegion

Translates this clipping region by the given vector.