Struct gfx::display_list::DisplayItemMetadata

pub struct DisplayItemMetadata {
    pub node: OpaqueNode,
    pub pointing: Option<Cursor>,
}

Metadata attached to each display item. This is useful for performing auxiliary tasks with the display list involving hit testing: finding the originating DOM node and determining the cursor to use when the element is hovered over.

Fields

node

The DOM node from which this display item originated.

pointing

The value of the cursor property when the mouse hovers over this display item. If None, this display item is ineligible for pointer events (pointer-events: none).

Methods

impl DisplayItemMetadata

fn new(node: OpaqueNode, style: &ComputedValues, default_cursor: Cursor) -> DisplayItemMetadata

Creates a new set of display metadata for a display item constributed by a DOM node. default_cursor specifies the cursor to use if cursor is auto. Typically, this will be PointerCursor, but for text display items it may be TextCursor or VerticalTextCursor.