Struct gfx::display_list::DisplayList

pub struct DisplayList {
    pub list: Vec<DisplayItem>,
    pub offsets: FnvHashMap<StackingContextId, StackingContextOffsets>,
    pub root_stacking_context: StackingContext,
}

Fields
list: Vec<DisplayItem> offsets: FnvHashMap<StackingContextId, StackingContextOffsets> root_stacking_context: StackingContext
Methods

impl DisplayList

fn new(root_stacking_context: StackingContext, items: &mut Option<Vec<DisplayItem>>) -> DisplayList

fn get_offset_for_item(&self, item: &DisplayItem) -> u32

fn sort(&mut self)

fn print(&self)

fn sort_and_count_stacking_contexts(stacking_context: &mut StackingContext, offsets: &mut HashMap<StackingContextId, StackingContextOffsets, BuildHasherDefault<FnvHasher>>, current_offset: u32) -> u32

fn print_with_tree(&self, print_tree: &mut PrintTree)

fn draw_item_at_index_into_context(&self, paint_context: &mut PaintContext, transform: &Matrix4D<f32>, index: usize)

Draws a single DisplayItem into the given PaintContext.

fn find_stacking_context<'a>(&'a self, stacking_context_id: StackingContextId) -> Option<&'a StackingContext>

fn draw_into_context<'a>(&self, paint_context: &mut PaintContext, transform: &Matrix4D<f32>, stacking_context_id: StackingContextId, start: usize, end: usize)

Draws the DisplayList in order.

fn draw_stacking_context_contents<'a>(&'a self, stacking_context: &StackingContext, traversal: &mut DisplayListTraversal<'a>, paint_context: &mut PaintContext, transform: &Matrix4D<f32>, tile_rect: Option<Rect<Au>>)

fn draw_stacking_context<'a>(&'a self, stacking_context: &StackingContext, traversal: &mut DisplayListTraversal<'a>, paint_context: &mut PaintContext, transform: &Matrix4D<f32>)

fn hit_test(&self, point: &Point2D<Au>, scroll_offsets: &ScrollOffsetMap) -> Vec<DisplayItemMetadata>

Places all nodes containing the point of interest into result, topmost first. Respects the pointer-events CSS property If topmost_only is true, stops after placing one node into the list. result must be empty upon entry to this function.