Struct gfx::display_list::ImageDisplayItem

pub struct ImageDisplayItem {
    pub base: BaseDisplayItem,
    pub image: Arc<Box<Image>>,
    pub stretch_size: Size2D<Au>,
    pub image_rendering: T,
}

Paints an image.

Fields

base

image

stretch_size

The dimensions to which the image display item should be stretched. If this is smaller than the bounds of this display item, then the image will be repeated in the appropriate direction to tile the entire bounds.

image_rendering

The algorithm we should use to stretch the image. See image_rendering in CSS-IMAGES-3  5.3.