Module gfx::filters

CSS and SVG filter support.

Functions

calculate_accumulated_blur

create_filters
Creates a filter pipeline from a set of CSS filters. Returns the destination end of the filter pipeline and the opacity.

grayscale
Creates a grayscale 5x4 color matrix per CSS-FILTERS  12.1.1.

hue_rotate
Creates a 5x4 hue rotation color matrix per CSS-FILTERS  8.5.

saturate
Creates a 5x4 saturation color matrix per CSS-FILTERS  8.5.

sepia
Creates a sepia 5x4 color matrix per CSS-FILTERS  12.1.1.

temporary_draw_target_needed_for_style_filters
Determines if we need a temporary draw target for the given set of filters.