Module gfx::font_cache_thread

Structs

FontCache 🔒

The font cache thread itself. It maintains a list of reference counted font templates that are currently in use.

FontCacheThread

The public interface to the font cache thread, used by per-thread FontContext instances (via the FontSource trait), and also by layout.

FontTemplateInfo

FontTemplates

A list of font templates that make up a given font family.

LowercaseString

SerializedFontTemplate

SerializedFontTemplateInfo

Enums

Command

Commands that the FontContext sends to the font cache thread.

FontIdentifier

Reply

Reply messages sent from the font cache thread to the FontContext caller.

Functions

is_supported_web_font_source 🔒

populate_generic_fonts 🔒