Struct gfx::font_cache_thread::FontTemplates

struct FontTemplates {
    templates: Vec<FontTemplate>,
}

A list of font templates that make up a given font family.

Fields

templates

Methods

impl FontTemplates

fn new() -> FontTemplates

fn find_font_for_style(&mut self, desc: &FontTemplateDescriptor, fctx: &FontContextHandle) -> Option<Arc<FontTemplateData>>

Find a font in this family that matches a given descriptor.

fn add_template(&mut self, identifier: Atom, maybe_data: Option<Vec<u8>>)