Struct gfx::font_context::FontContext

pub struct FontContext {
    platform_handle: FontContextHandle,
    font_cache_thread: FontCacheThread,
    layout_font_cache: Vec<LayoutFontCacheEntry>,
    fallback_font_cache: Vec<FallbackFontCacheEntry>,
    layout_font_group_cache: HashMap<LayoutFontGroupCacheKey, Rc<FontGroup>, BuildHasherDefault<FnvHasher>>,
    epoch: usize,
}

The FontContext represents the per-thread/thread state necessary for working with fonts. It is the public API used by the layout and paint code. It talks directly to the font cache thread where required.

Fields
platform_handle: FontContextHandle font_cache_thread: FontCacheThread layout_font_cache: Vec<LayoutFontCacheEntry>

TODO: See bug https://github.com/servo/servo/issues/3300.

fallback_font_cache: Vec<FallbackFontCacheEntry> layout_font_group_cache: HashMap<LayoutFontGroupCacheKey, Rc<FontGroup>, BuildHasherDefault<FnvHasher>> epoch: usize
Methods

impl FontContext

fn new(font_cache_thread: FontCacheThread) -> FontContext

fn create_layout_font(&self,
template: Arc<FontTemplateData>,
descriptor: FontTemplateDescriptor,
pt_size: Au,
variant: T,
font_key: FontKey)
-> Result<Font, ()>

Create a font for use in layout calculations.

fn expire_font_caches_if_necessary(&mut self)

fn layout_font_group_for_style(&mut self, style: Arc<Font>) -> Rc<FontGroup>

Create a group of fonts for use in layout calculations. May return a cached font if this font instance has already been used by this context.