Module gfx::font_template

Structs

FontTemplate
This describes all the information needed to create font instance handles. It contains a unique FontTemplateData structure that is platform specific.

FontTemplateDescriptor
Describes how to select a font from a given family. This is very basic at the moment and needs to be expanded or refactored when we support more of the font styling parameters.

Functions

style_to_number