Struct gfx::font_template::FontTemplateDescriptor

pub struct FontTemplateDescriptor {
    pub weight: T,
    pub stretch: T,
    pub italic: bool,
}

Describes how to select a font from a given family. This is very basic at the moment and needs to be expanded or refactored when we support more of the font styling parameters.

Fields

weight

stretch

italic

Methods

impl FontTemplateDescriptor

fn new(weight: T, stretch: T, italic: bool) -> FontTemplateDescriptor