Module gfx::handle

Resource handles

Structs

Buffer
Type-safe buffer handle

DepthStencilView
Typed DSV

Fence
Fence Handle

Manager
Stores reference-counted resources used in a command buffer. Seals actual resource names behind the interface, automatically referencing them both by the Factory on resource creation and the Renderer during CommandBuffer population.

Program
Program Handle

RawBuffer
Untyped buffer handle

RawDepthStencilView
Raw DSV

RawPipelineState
Raw Pipeline State Handle

RawRenderTargetView
Raw RTV

RawShaderResourceView
Raw Shader Resource View Handle

RawTexture
Raw texture handle

RawUnorderedAccessView
Raw Unordered Access View Handle

RenderTargetView
Typed RTV

Sampler
Sampler Handle

Shader
Shader Handle

ShaderResourceView
Type-safe Shader Resource View Handle

Texture
Typed texture object

UnorderedAccessView
Type-safe Unordered Access View Handle