Struct gfx::paint_context::PaintContext

pub struct PaintContext<'a> {
    pub draw_target: DrawTarget,
    pub font_context: &'a mut Box<FontContext>,
    pub page_rect: TypedRect<PagePx, f32>,
    pub screen_rect: TypedRect<ScreenPx, usize>,
    pub clip_rect: Option<Rect<Au>>,
    pub transient_clip: Option<ClippingRegion>,
    pub layer_kind: LayerKind,
}

Fields

draw_target

font_context

page_rect

The rectangle that this context encompasses in page coordinates.

screen_rect

The rectangle that this context encompasses in screen coordinates (pixels).

clip_rect

The clipping rect for the stacking context as a whole.

transient_clip

The current transient clipping region, if any. A "transient clipping region" is the clipping region used by the last display item. We cache the last value so that we avoid pushing and popping clipping regions unnecessarily.

layer_kind

A temporary hack to disable clipping optimizations on 3d layers.

Methods

impl<'a> PaintContext<'a>

fn screen_pixels_per_px(&self) -> ScaleFactor<PagePx, ScreenPx, f32>

fn draw_target(&self) -> &DrawTarget

fn draw_solid_color(&self, bounds: &Rect<Au>, color: Color)

fn draw_border(&self, bounds: &Rect<Au>, border: &SideOffsets2D<Au>, radius: &BorderRadii<Au>, color: &SideOffsets2D<Color>, style: &SideOffsets2D<T>)

fn draw_line(&self, bounds: &Rect<Au>, color: Color, style: T)

fn draw_push_clip(&self, bounds: &Rect<Au>)

fn draw_pop_clip(&self)

fn draw_image(&self, bounds: &Rect<Au>, stretch_size: &Size2D<Au>, image_info: &WebRenderImageInfo, image_data: &[u8], image_rendering: T)

fn clear(&self)

fn draw_border_segment(&self, direction: Direction, bounds: &Rect<Au>, border: &SideOffsets2D<f32>, radius: &BorderRadii<AzFloat>, color: &SideOffsets2D<Color>, style: &SideOffsets2D<T>)

fn draw_line_segment(&self, bounds: &Rect<Au>, radius: &BorderRadii<AzFloat>, color: Color, style: T)

fn draw_border_path(&self, bounds: &Rect<f32>, direction: Direction, border: &SideOffsets2D<f32>, radii: &BorderRadii<AzFloat>, color: Color)

fn push_rounded_rect_clip(&self, bounds: &Rect<f32>, radii: &BorderRadii<AzFloat>)

fn solve_quadratic(a: f32, b: f32, c: f32) -> (Option<f32>, Option<f32>)

fn intersect_ellipse_line(e: Ellipse, line: Line) -> (Option<Point2D<f32>>, Option<Point2D<f32>>)

fn ellipse_line_intersection_angles(e: Ellipse, l: Line) -> (Option<(Point2D<f32>, f32)>, Option<(Point2D<f32>, f32)>)

fn ellipse_rightmost_intersection(e: Ellipse, l: Line) -> Option<f32>

fn ellipse_leftmost_intersection(e: Ellipse, l: Line) -> Option<f32>

fn is_zero_radius(radius: &Size2D<AzFloat>) -> bool

fn ellipse_to_bezier(path_builder: &mut PathBuilder, origin: Point2D<AzFloat>, radius: Size2D<AzFloat>, start_angle: f32, end_angle: f32)

fn inner_border_bounds(bounds: &Rect<f32>, border: &SideOffsets2D<f32>) -> Rect<f32>

fn corner_bounds(bounds: &Rect<f32>, border: &SideOffsets2D<f32>, radii: &BorderRadii<AzFloat>) -> (CornerOrigin, SideOffsets2D<Size2D<f32>>)

fn draw_corner(path_builder: &mut PathBuilder, corner: BorderCorner, origin: &Point2D<AzFloat>, radius: &Size2D<AzFloat>, inner_border: &Point2D<AzFloat>, outer_border: &Point2D<AzFloat>, dist_elbow: &Size2D<AzFloat>, clockwise: bool)

origin is the origin point when drawing the corner e.g. it's the circle center when drawing radial borders.

corner indicates which corner to draw e.g. top left or top right etc.

radius is the border-radius width and height. If radius.width == radius.height then an arc from a circle is drawn instead of an arc from an ellipse.

inner_border & outer_border are the inner and outer points on the border corner respectively. ASCII diagram: ---------------* =====> ("*" is the outer_border point) | | | --------* ============> ("*" is the inner_border point) | | | |

dist_elbow is the distance from origin to the inner part of the border corner. clockwise indicates direction to draw the border curve.

fn create_border_path_segment(&self, path_builder: &mut PathBuilder, bounds: &Rect<f32>, direction: Direction, border: &SideOffsets2D<f32>, radii: &BorderRadii<AzFloat>, mode: BorderPathDrawingMode)

fn create_rounded_rect_path(&self, path_builder: &mut PathBuilder, bounds: &Rect<f32>, radii: &BorderRadii<AzFloat>)

Creates a path representing the given rounded rectangle.

TODO(pcwalton): Should we unify with the code above? It doesn't seem immediately obvious how to do that (especially without regressing performance) unless we have some way to efficiently intersect or union paths, since different border styles/colors can force us to slice through the rounded corners. My first attempt to unify with the above code resulted in making a mess of it, and the simplicity of this code path is appealing, so it may not be worth it… In any case, revisit this decision when we support elliptical radii.

fn draw_dashed_border_segment(&self, direction: Direction, bounds: &Rect<Au>, border: &SideOffsets2D<f32>, radius: &BorderRadii<AzFloat>, color: Color, dash_size: DashSize)

fn draw_solid_border_segment(&self, direction: Direction, bounds: &Rect<Au>, border: &SideOffsets2D<f32>, radius: &BorderRadii<AzFloat>, color: Color)

fn compute_scaled_bounds(&self, bounds: &Rect<Au>, border: &SideOffsets2D<f32>, shrink_factor: f32) -> Rect<f32>

fn scale_color(&self, color: Color, scale_factor: f32) -> Color

fn draw_double_border_segment(&self, direction: Direction, bounds: &Rect<Au>, border: &SideOffsets2D<f32>, radius: &BorderRadii<AzFloat>, color: Color)

fn draw_groove_ridge_border_segment(&self, direction: Direction, bounds: &Rect<Au>, border: &SideOffsets2D<f32>, radius: &BorderRadii<AzFloat>, color: Color, style: T)

fn draw_inset_outset_border_segment(&self, direction: Direction, bounds: &Rect<Au>, border: &SideOffsets2D<f32>, radius: &BorderRadii<AzFloat>, color: Color, style: T)

fn draw_text(&mut self, text: &TextDisplayItem)

Draws the given text display item into the current context.

fn draw_linear_gradient(&self, bounds: &Rect<Au>, start_point: &Point2D<Au>, end_point: &Point2D<Au>, stops: &[GradientStop])

Draws a linear gradient in the given boundaries from the given start point to the given end point with the given stops.

fn get_or_create_temporary_draw_target(&mut self, filters: &T, blend_mode: T) -> DrawTarget

fn draw_temporary_draw_target_if_necessary(&mut self, temporary_draw_target: &DrawTarget, filters: &T, blend_mode: T)

If we created a temporary draw target, then draw it to the main draw target. This is called after doing all the painting, and the temporary draw target must not be used afterward.

fn draw_box_shadow(&mut self, box_bounds: &Rect<Au>, offset: &Point2D<Au>, color: Color, blur_radius: Au, spread_radius: Au, clip_mode: BoxShadowClipMode)

Draws a box shadow with the given boundaries, color, offset, blur radius, and spread radius. box_bounds represents the boundaries of the box.

fn create_draw_target_for_blur_if_necessary(&self, box_bounds: &Rect<Au>, blur_radius: Au) -> TemporaryDrawTarget

If we have blur, create a new draw target that's the same size as this tile, but with enough space around the edges to hold the entire blur. (If we don't do the latter, then there will be seams between tiles.)

fn blur_if_necessary(&self, temporary_draw_target: TemporaryDrawTarget, blur_radius: Au)

Performs a blur using the draw target created in create_draw_target_for_blur_if_necessary.

fn push_clip_if_applicable(&self)

fn pop_clip_if_applicable(&self)

fn remove_transient_clip_if_applicable(&mut self)

fn push_transient_clip(&mut self, clip_region: ClippingRegion)

Sets a new transient clipping region. Automatically calls remove_transient_clip_if_applicable() first.