Struct gfx::paint_context::TemporaryDrawTarget

struct TemporaryDrawTarget {
    draw_target: DrawTarget,
    offset: Point2D<AzFloat>,
}

Represents a temporary drawing surface. Some operations that perform complex compositing operations need this.

Fields
draw_target: DrawTarget

The draw target.

offset: Point2D<AzFloat>

The distance from the top left of the main draw target to the top left of this temporary draw target.

Methods

impl TemporaryDrawTarget

fn from_main_draw_target(main_draw_target: &DrawTarget) -> TemporaryDrawTarget

Creates a temporary draw target that simply draws to the main draw target.

fn from_bounds(main_draw_target: &DrawTarget, bounds: &Rect<Au>) -> TemporaryDrawTarget

Creates a temporary draw target large enough to encompass the given bounding rect in page coordinates. The temporary draw target will have the same transform as the tile we're drawing to.

fn draw_filter(self, main_draw_target: &DrawTarget, filter: FilterNode)

Composites this temporary draw target onto the main surface, with the given Azure filter.