Struct gfx::paint_thread::PaintLayer

struct PaintLayer {
    pub layer_properties: LayerProperties,
    pub starting_stacking_context_id: StackingContextId,
    pub display_list_indices: Option<(usize, usize)>,
    pub single_item: bool,
    pub display_list_origin: Point2D<f32>,
}

Fields

layer_properties

The LayerProperties, which describe the layer in a way that the Compositor can consume.

starting_stacking_context_id

The StackingContextId of the StackingContext that is the immediate parent of this layer. This is used to ensure applying the proper transform when painting.

display_list_indices

The indices (in the DisplayList) to the first and last display item that are the contents of this layer.

single_item

When painting, whether to draw the start by entering the surrounding StackingContext or simply to draw the single item this PaintLayer contains.

display_list_origin

The layer's bounds start at the overflow origin, but display items are positioned relative to the stacking context bounds, so we need to offset by the overflow rect (which will be in the coordinate system of the stacking context bounds).

Methods

impl PaintLayer

fn new_from_stacking_context(layer_info: &LayerInfo, stacking_context: &StackingContext, parent_origin: &Point2D<Au>, transform: &Matrix4D<f32>, perspective: &Matrix4D<f32>, parent_id: Option<LayerId>) -> PaintLayer

fn new_for_display_item(layer_info: &LayerInfo, item_bounds: &Rect<Au>, parent_origin: &Point2D<Au>, transform: &Matrix4D<f32>, perspective: &Matrix4D<f32>, parent_id: Option<LayerId>, stacking_context_id: StackingContextId, item_index: usize) -> PaintLayer

fn add_item(&mut self, index: usize)

fn make_companion_layer(&mut self)