Struct gfx::paint_thread::PaintThread

pub struct PaintThread<C> {
    id: PipelineId,
    _url: Url,
    layout_to_paint_port: Receiver<LayoutToPaintMsg>,
    chrome_to_paint_port: Receiver<ChromeToPaintMsg>,
    compositor: C,
    time_profiler_chan: ProfilerChan,
    root_display_list: Option<Arc<DisplayList>>,
    layer_map: HashMap<LayerId, Arc<PaintLayer>>,
    paint_permission: bool,
    current_epoch: Option<Epoch>,
    worker_threads: Vec<WorkerThreadProxy>,
}

Fields

id

_url

layout_to_paint_port

chrome_to_paint_port

compositor

time_profiler_chan

A channel to the time profiler.

root_display_list

The root paint layer sent to us by the layout thread.

layer_map

A map that associates LayerIds with their corresponding layers.

paint_permission

Permission to send paint messages to the compositor

current_epoch

The current epoch counter is passed by the layout thread

worker_threads

Communication handles to each of the worker threads.

Methods

impl<C> PaintThread<C> where C: PaintListener + Send + 'static

fn create(id: PipelineId, url: Url, chrome_to_paint_chan: Sender<ChromeToPaintMsg>, layout_to_paint_port: Receiver<LayoutToPaintMsg>, chrome_to_paint_port: Receiver<ChromeToPaintMsg>, compositor: C, panic_chan: IpcSender<PanicMsg>, font_cache_thread: FontCacheThread, time_profiler_chan: ProfilerChan, mem_profiler_chan: ProfilerChan, shutdown_chan: IpcSender<()>)

fn start(&mut self)

fn paint(&mut self, replies: &mut Vec<(LayerId, Box<LayerBufferSet>)>, tiles: Vec<BufferRequest>, scale: f32, layer_id: LayerId, layer_kind: LayerKind)

Paints one layer and places the painted tiles in replies.

fn initialize_layers(&mut self)