Struct gfx::platform::freetype::font_template::FontTemplateData

pub struct FontTemplateData {
    pub bytes: Vec<u8>,
    pub identifier: Atom,
}

Platform specific font representation for Linux. The identifier is an absolute path, and the bytes field is the loaded data that can be passed to freetype and azure directly.

Fields
bytes: Vec<u8> identifier: Atom
Methods

impl FontTemplateData

fn new(identifier: Atom,
font_data: Option<Vec<u8>>)
-> Result<FontTemplateData, Error>

fn bytes(&self) -> Vec<u8>

Returns a clone of the data in this font. This may be a hugely expensive operation (depending on the platform) which performs synchronous disk I/O and should never be done lightly.

fn bytes_if_in_memory(&self) -> Option<Vec<u8>>

Returns a clone of the bytes in this font if they are in memory. This function never performs disk I/O.

fn native_font(&self) -> Option<NativeFontHandle>

Returns the native font that underlies this font template, if applicable.