Module gfx::pso::resource

Resource components for a PSO.

Structs

RawShaderResource
Raw (untyped) shader resource (SRV).

Sampler
Sampler component.

ShaderResource
Shader resource component (SRV). Typically is a view into some texture, but can also be a buffer.

TextureSampler
A convenience type for a texture paired with a sampler. It only makes sense for DX9 class hardware, where every texture by default is bundled with a sampler, hence they are represented by the same name. In DX10 and higher samplers are totally separated from the textures.

UnorderedAccess
Unordered access component (UAV). A writable resource (texture/buffer) with no defined access order across simultaneously executing shaders. Supported on DX10 and higher.