Struct gfx::pso::resource::TextureSampler

pub struct TextureSampler<T>(_, _);

A convenience type for a texture paired with a sampler. It only makes sense for DX9 class hardware, where every texture by default is bundled with a sampler, hence they are represented by the same name. In DX10 and higher samplers are totally separated from the textures.

init: &str = name of the sampler/texture (assuming they match)

data: (ShaderResourceView<T>, Sampler)