Module gfx::render

High-level, platform independent, bindless rendering API.

Modules

batch
Batches Batches are structures containing all the data required for the draw call, except for the target frame. Here we define the Batch trait as well as RefBatch and OwnedBatch implementations.

mesh
Meshes Mesh loading.

shade
Shaders Shader parameter handling.

target
Render targets Render target specification.

Structs

ParamStorage
Temporary parameter storage, used for shader activation.

Renderer
Renderer front-end

Enums

BlitError
An error occuring in surface blits.

DrawError
Draw-time error, showing inconsistencies in draw parameters and data

UpdateError
An error occuring in buffer/texture updates.

Traits

RenderFactory
Factory extension that allows creating new renderers.