Module gfx::render::batch

Batches Batches are structures containing all the data required for the draw call, except for the target frame. Here we define the Batch trait as well as RefBatch and OwnedBatch implementations.

Structs

Core
Core batch - a minimal sealed batch.

Full
Full batch - contains everything needed for rendering.

Enums

Error
An error occurring at batch creation

Traits

Batch
Abstract batch trait

Functions

bind
Create an implicit batch

Type Definitions

BatchData
Return type for Batch::get_data()`

Complete
A Core completed by a mesh slice, shader parameters, and a state. Implements Batch thus can be drawn. It is meant to be a struct, but we have lots of lifetime issues with associated resources, binding which looks nasty (#614)

Implicit
A batch that is constructed on the fly when rendering. Meant to be a struct, blocked by #614