Module gfx::render::mesh

Meshes Mesh loading.

A Mesh describes the geometry of an object. All or part of a Mesh can be drawn with a single draw call. A Mesh consists of a series of vertices, each with a certain amount of user-specified attributes. These attributes are fed into shader programs. The easiest way to create a mesh is to use the #[vertex_format] attribute on a struct, upload them into a Buffer, and then use Mesh::from.

Structs

Attribute
Describes a single attribute of a vertex buffer, including its type, name, etc.

AttributeIter
An iterator over mesh attributes.

Link
Holds a remapping table from shader inputs to mesh attributes.

Mesh
Describes geometry to render.

Slice
Description of a subset of Mesh data to render.

Enums

Error
Describes kinds of errors that may occur in the mesh linking

SliceKind
Source of vertex ordering for a slice

Traits

ToIndexSlice
A helper trait to build index slices from data.

ToSlice
A helper trait for cleanly getting the slice of a type.

VertexFormat
A trait implemented automatically for user vertex structure by `#[vertex_format] attribute

Type Definitions

AttributeIndex
Index of a vertex attribute inside the mesh