Struct gfx::render::mesh::Link

pub struct Link {
    // some fields omitted
}

Holds a remapping table from shader inputs to mesh attributes.

Methods

impl Link

fn new<R: Resources>(mesh: &Mesh<R>, pinfo: &ProgramInfo) -> Result<Link, Error>

Match mesh attributes against shader inputs, produce a mesh link. Exposed to public to allow external Batch implementations to use it.

fn from_iter<I: Iterator<Item=AttributeIndex>>(iter: I) -> Result<Link, Error>

Construct a new link from an iterator over attribute indices.

fn to_iter(&self) -> AttributeIter

Convert to an iterator returning attribute indices