Module gfx::render::shade

Shaders Shader parameter handling.

Reexports

pub use device::shade::{Stage, CreateShaderError};

Structs

NamedCell
A named cell containing arbitrary value

ParamDictionary
A dictionary of parameters, meant to be shared between different programs

ParamDictionaryLink
Redirects program input to the relevant ParamDictionary cell

Enums

ParameterError
An error type on either the parameter storage or the program side

Traits

ShaderParam
Abstracts the shader parameter structure, generated by the shader_param attribute

ToUniform
Helper trait to transform base types into their corresponding uniforms

Type Definitions

TextureParam
A texture parameter: consists of a texture handle with an optional sampler.

VarBlock
Variable index of a uniform block.

VarTexture
Variable index of a texture.

VarUniform
Variable index of a uniform.