Trait gfx::render::shade::Parameter

pub trait Parameter<R: Resources> {
    fn put(&self, ParameterId, &mut ParamStorage<R>);

    fn check_uniform(&UniformVar) -> bool { ... }
    fn check_block(&BlockVar) -> bool { ... }
    fn check_texture(&SamplerVar) -> bool { ... }
}

General shader parameter.

Required Methods

fn put(&self, ParameterId, &mut ParamStorage<R>)

Write into the parameter storage for rendering.

Provided Methods

fn check_uniform(&UniformVar) -> bool

Check if this parameter is good for a given uniform.

fn check_block(&BlockVar) -> bool

Check if this parameter is good for a given block.

fn check_texture(&SamplerVar) -> bool

Check if this parameter is good for a given texture.