gfx::render::shade::TextureParam

type TextureParam<R: Resources> = (Texture<R>, Option<Sampler<R>>);

A texture parameter: consists of a texture handle with an optional sampler. Not all textures need a sampler (i.e. MSAA ones do not). Optimally, we'd want to encode this logic into the type system (TODO).