Enum gfx::render::target::Plane

pub enum Plane<R: Resources> {
    Surface(Surface<R>),
    Texture(Texture<R>, Level, Option<Layer>),
}

A single buffer that can be bound to a render target.

Variants

Surface

Render to a Surface (corresponds to a renderbuffer in GL).

Texture

Render to a texture at a specific mipmap level If Layer is set, it is selecting a single 2D slice of a given 3D texture

Methods

impl<R: Resources> Plane<R>

fn get_surface_info(&self) -> SurfaceInfo

Get the surface info.

fn get_format(&self) -> Format

Get surface/texture format.