Struct gfx::render::target::Frame

pub struct Frame<R: Resources> {
    pub width: u16,
    pub height: u16,
    pub colors: Vec<Plane<R>>,
    pub depth: Option<Plane<R>>,
    pub stencil: Option<Plane<R>>,
    pub convert_gamma: bool,
}

A complete Frame, which is the result of rendering.

Fields

width

The width of the viewport.

height

The height of the viewport.

colors

Each color component has its own buffer.

depth

The depth buffer for this frame.

stencil

The stencil buffer for this frame.

convert_gamma

Convert to sRGB color space.

Methods

impl<R: Resources> Frame<R>

fn new(width: u16, height: u16) -> Frame<R>

Create an empty Frame, which corresponds to the 'default framebuffer', which renders directly to the window that was created with the OpenGL context.

fn is_default(&self) -> bool

Return true if this framebuffer is associated with the main window (matches Frame::new exactly).

fn get_mask(&self) -> Mask

Return a mask of contained planes.