Module gfx::shade::core

Shader handling.

Structs

AttributeVar

Vertex information that a shader takes as input.

ConstVar

A constant in the shader - a bit of data that doesn’t vary

ConstantBufferVar

A constant buffer.

CreateProgramError

An error type for creating programs.

OutputVar

Target output variable.

Program

A program

ProgramInfo

Metadata about a program.

SamplerType

A type of the sampler variable.

SamplerVar

Sampler shader parameter.

TextureVar

Texture shader parameter.

UnorderedVar

Unordered access shader parameter.

Usage

Parameter usage flags.

Enums

BaseType

Base type of this shader parameter.

CompatibilityError

Error type for trying to store a UniformValue in a ConstVar.

ContainerType

Number of components this parameter represents.

CreateShaderError

An error type for creating shaders.

IsArray

Whether the sampler samples an array texture.

IsComparison

Whether the sampler compares the depth value upon sampling.

IsMultiSample

Whether the sampler samples a multisample texture.

IsRect

Whether the sampler samples a rectangle texture.

MatrixFormat

Whether the matrix is column or row major.

Stage

Which program stage this shader represents.

TextureType

A type of the texture variable. This has to match the actual data we bind to the shader.

UniformValue

A value that can be uploaded to the device as a uniform.

Constants

STAGES

A constant static array of all shader stages.

Traits

BaseTyped

A trait that statically links simple data types to base types of the shader constants.

Formatted

A trait that statically links simple data types to constant formats.

Type Definitions

ConstFormat

Format of a shader constant.

Dimension

Number of components in a container type (vectors/matrices)

Location

Location of a parameter in the program.