Struct gfx::state::StencilSide

pub struct StencilSide {
    pub fun: Comparison,
    pub value: u8,
    pub mask_read: u8,
    pub mask_write: u8,
    pub op_fail: StencilOp,
    pub op_depth_fail: StencilOp,
    pub op_pass: StencilOp,
}

Complete stencil state for a given side of a face.

Fields

fun

Comparison function to use to determine if the stencil test passes.

value

Reference value to compare the value in the stencil buffer with.

mask_read

A mask that is ANDd with both the stencil buffer value and the reference value when they are read before doing the stencil test.

mask_write

This is unused!

op_fail

What operation to do if the stencil test fails.

op_depth_fail

What operation to do if the stenil test passes but the depth test fails.

op_pass

What operation to do if both the depth and stencil test pass.