Enum gfx::state::StencilOp

pub enum StencilOp {
    OpKeep,
    OpZero,
    OpReplace,
    OpIncrementClamp,
    OpIncrementWrap,
    OpDecrementClamp,
    OpDecrementWrap,
    OpInvert,
}

Stencil mask operation.

Variants

OpKeep

Keep the current value in the stencil buffer (no change).

OpZero

Set the value in the stencil buffer to zero.

OpReplace

Set the stencil buffer value to value from StencilSide

OpIncrementClamp

Increment the stencil buffer value, clamping to its maximum value.

OpIncrementWrap

Increment the stencil buffer value, wrapping around to 0 on overflow.

OpDecrementClamp

Decrement the stencil buffer value, clamping to its minimum value.

OpDecrementWrap

Decrement the stencil buffer value, wrapping around to the maximum value on overflow.

OpInvert

Bitwise invert the current value in the stencil buffer.