Struct gfx::Frame

pub struct Frame {
    pub width: u16,
    pub height: u16,
    pub colors: [Plane, ..4],
    pub depth: Plane,
    pub stencil: Plane,
}

A complete Frame, which is the result of rendering.

Fields

width

The width of the viewport.

height

The height of the viewport.

colors

Each color component has its own buffer.

depth

The depth buffer for this frame.

stencil

The stencil buffer for this frame.

Methods

impl Frame

fn new(width: u16, height: u16) -> Frame

Create an empty Frame, which corresponds to the 'default framebuffer', which for now renders directly to the window that was created with the OpenGL context.

fn is_default(&self) -> bool

Returns true if this framebuffer is associated with the main window (matches Frame::new exactly).