Struct gfx::PaintContext

pub struct PaintContext<'a> {
    pub draw_target: DrawTarget,
    pub font_context: &'a mut Box<FontContext>,
    pub page_rect: Rect<f32>,
    pub screen_rect: Rect<usize>,
    pub clip_rect: Option<Rect<Au>>,
    pub transient_clip: Option<ClippingRegion>,
}

Fields

draw_target

font_context

page_rect

The rectangle that this context encompasses in page coordinates.

screen_rect

The rectangle that this context encompasses in screen coordinates (pixels).

clip_rect

The clipping rect for the stacking context as a whole.

transient_clip

The current transient clipping region, if any. A "transient clipping region" is the clipping region used by the last display item. We cache the last value so that we avoid pushing and popping clipping regions unnecessarily.

Methods

impl<'a> PaintContext<'a>

fn get_draw_target(&self) -> &DrawTarget

fn draw_solid_color(&self, bounds: &Rect<Au>, color: Color)

fn draw_border(&self, bounds: &Rect<Au>, border: &SideOffsets2D<Au>, radius: &BorderRadii<Au>, color: &SideOffsets2D<Color>, style: &SideOffsets2D<T>)

fn draw_line(&self, bounds: &Rect<Au>, color: Color, style: T)

fn draw_push_clip(&self, bounds: &Rect<Au>)

fn draw_pop_clip(&self)

fn draw_image(&self, bounds: &Rect<Au>, image: Arc<Box<Image>>, image_rendering: T)

fn clear(&self)

fn draw_text(&mut self, text: &TextDisplayItem)

Draws the given text display item into the current context.

fn draw_linear_gradient(&self, bounds: &Rect<Au>, start_point: &Point2D<Au>, end_point: &Point2D<Au>, stops: &[GradientStop])

Draws a linear gradient in the given boundaries from the given start point to the given end point with the given stops.

fn get_or_create_temporary_draw_target(&mut self, filters: &T, blend_mode: T) -> DrawTarget

fn draw_temporary_draw_target_if_necessary(&mut self, temporary_draw_target: &DrawTarget, filters: &T, blend_mode: T)

If we created a temporary draw target, then draw it to the main draw target. This is called after doing all the painting, and the temporary draw target must not be used afterward.

fn draw_box_shadow(&mut self, box_bounds: &Rect<Au>, offset: &Point2D<Au>, color: Color, blur_radius: Au, spread_radius: Au, clip_mode: BoxShadowClipMode)

Draws a box shadow with the given boundaries, color, offset, blur radius, and spread radius. box_bounds represents the boundaries of the box.

fn push_clip_if_applicable(&self)

fn pop_clip_if_applicable(&self)

fn remove_transient_clip_if_applicable(&mut self)

fn push_transient_clip(&mut self, clip_region: ClippingRegion)

Sets a new transient clipping region. Automatically calls remove_transient_clip_if_applicable() first.