Struct gfx::ProgramInfo

pub struct ProgramInfo {
    pub vertex_attributes: Vec<AttributeVar>,
    pub globals: Vec<ConstVar>,
    pub constant_buffers: Vec<ConstantBufferVar>,
    pub textures: Vec<TextureVar>,
    pub unordereds: Vec<UnorderedVar>,
    pub samplers: Vec<SamplerVar>,
    pub outputs: Vec<OutputVar>,
    pub output_depth: bool,
    pub knows_outputs: bool,
}

Metadata about a program.

Fields
vertex_attributes: Vec<AttributeVar>

Attributes in the program

globals: Vec<ConstVar>

Global constants in the program

constant_buffers: Vec<ConstantBufferVar>

Constant buffers in the program

textures: Vec<TextureVar>

Textures in the program

unordereds: Vec<UnorderedVar>

Unordered access resources in the program

samplers: Vec<SamplerVar>

Samplers in the program

outputs: Vec<OutputVar>

Output targets in the program

output_depth: bool

A flag indicating that the pixel shader manually assigns the depth.

knows_outputs: bool

A hacky flag to make sure the clients know we are unable to actually get the output variable info