Enum gfx::tex::Format

pub enum Format {
    Float(Components, FloatSize),
    Integer(Components, u8, IntSubType),
    Unsigned(Components, u8, IntSubType),
    R3G3B2,
    RGB5A1,
    RGB10A2,
    RGB10A2UI,
    R11FG11FB10F,
    RGB9E5,
    DEPTH24STENCIL8,
}

Describes the layout of each texel within a surface/texture.

Variants

Float

Floating point.

Integer

Signed integer.

Unsigned

Unsigned integer.

R3G3B2

Normalized integer, with 3 bits for R and G, but only 2 for B.

RGB5A1

5 bits each for RGB, 1 for Alpha.

RGB10A2

10 bits each for RGB, 2 for Alpha.

RGB10A2UI

10 bits each for RGB, 2 for Alpha, as unsigned integers.

R11FG11FB10F

This uses special 11 and 10-bit floating-point values without sign bits.

RGB9E5

This s an RGB format of type floating-point. The 3 color values have 9 bits of precision, and they share a single exponent.

DEPTH24STENCIL8

24 bits for depth, 8 for stencil