Enum gfx::tex::TextureKind

pub enum TextureKind {
    Texture1D,
    Texture1DArray,
    Texture2D,
    Texture2DArray,
    TextureCube,
    Texture3D,
}

Specifies how a given texture may be used. The available texture types are restricted by what Metal exposes, though this could conceivably be extended in the future. Note that a single texture can only ever be of one kind. A texture created as Texture2D will forever be Texture2D.

Variants

Texture1D

A single row of texels.

Texture1DArray

An array of rows of texels. Equivalent to Texture2D except that texels in a different row are not sampled.

Texture2D

A traditional 2D texture, with rows arranged contiguously.

Texture2DArray

An array of 2D textures. Equivalent to Texture3D except that texels in a different depth level are not sampled.

TextureCube

A set of 6 2D textures, one for each face of a cube.

Texture3D

A volume texture, with each 2D layer arranged contiguously.