Module gfx::text::glyph

Structs

ByteIndex
An index that refers to a byte offset in a text run. This could point to the middle of a glyph.

DetailedGlyph

DetailedGlyphRecord

DetailedGlyphStore

GlyphData

GlyphEntry
GlyphEntry is a port of Gecko's CompressedGlyph scheme for storing glyph data compactly.

GlyphIterator
An iterator over the glyphs in a byte range in a GlyphStore.

GlyphStore
Stores the glyph data belonging to a text run.

Enums

GlyphInfo

Constants

FLAG_CHAR_IS_SPACE

FLAG_IS_SIMPLE_GLYPH

GLYPH_ADVANCE_MASK

GLYPH_ADVANCE_SHIFT

GLYPH_COUNT_MASK

GLYPH_ID_MASK

Functions

is_simple_advance

is_simple_glyph_id

Type Definitions

GlyphId
The id of a particular glyph within a font