Struct gfx::text::glyph::GlyphEntry

struct GlyphEntry {
    value: u32,
}

GlyphEntry is a port of Gecko's CompressedGlyph scheme for storing glyph data compactly.

In the common case (reasonable glyph advances, no offsets from the font em-box, and one glyph per character), we pack glyph advance, glyph id, and some flags into a single u32.

In the uncommon case (multiple glyphs per unicode character, large glyph index/advance, or glyph offsets), we pack the glyph count into GlyphEntry, and store the other glyph information in DetailedGlyphStore.

Fields

value

Methods

impl GlyphEntry

fn new(value: u32) -> GlyphEntry

fn initial() -> GlyphEntry

fn simple(id: GlyphId, advance: Au) -> GlyphEntry

fn complex(starts_cluster: bool, starts_ligature: bool, glyph_count: usize) -> GlyphEntry

fn is_initial(&self) -> bool

impl GlyphEntry

fn advance(&self) -> Au

fn id(&self) -> GlyphId

fn char_is_space(&self) -> bool

True if original char was normal (U+0020) space. Other chars may map to space glyph, but this does not account for them.

fn set_char_is_space(&mut self)

fn glyph_count(&self) -> u16

fn is_simple(&self) -> bool

fn has_flag(&self, flag: u32) -> bool