Struct gfx::text::glyph::GlyphStore

pub struct GlyphStore {
    entry_buffer: Vec<GlyphEntry>,
    detail_store: DetailedGlyphStore,
    total_advance: Au,
    total_spaces: i32,
    has_detailed_glyphs: bool,
    is_whitespace: bool,
    is_rtl: bool,
}

Stores the glyph data belonging to a text run.

Simple glyphs are stored inline in the entry_buffer, detailed glyphs are stored as pointers into the detail_store.

+- GlyphStore --------------------------------+
|               +---+---+---+---+---+---+---+ |
| entry_buffer: |   | s |   | s |   | s | s | |  d = detailed
|               +-|-+---+-|-+---+-|-+---+---+ |  s = simple
|                 |       |       |           |
|                 |   +---+-------+           |
|                 |   |                       |
|               +-V-+-V-+                     |
| detail_store: | d | d |                     |
|               +---+---+                     |
+---------------------------------------------+

Fields
entry_buffer: Vec<GlyphEntry>

A buffer of glyphs within the text run, in the order in which they appear in the input text. Any changes will also need to be reflected in transmute_entry_buffer_to_u32_buffer().

detail_store: DetailedGlyphStore

A store of the detailed glyph data. Detailed glyphs contained in the entry_buffer point to locations in this data structure.

total_advance: Au

A cache of the advance of the entire glyph store.

total_spaces: i32

A cache of the number of spaces in the entire glyph store.

has_detailed_glyphs: bool

Used to check if fast path should be used in glyph iteration.

is_whitespace: bool is_rtl: bool
Methods

impl<'a> GlyphStore

fn new(length: usize, is_whitespace: bool, is_rtl: bool) -> GlyphStore

Initializes the glyph store, but doesn't actually shape anything.

Use the add_* methods to store glyph data.

fn len(&self) -> ByteIndex

fn is_whitespace(&self) -> bool

fn finalize_changes(&mut self)

fn cache_total_advance_and_spaces(&mut self)

fn add_glyph_for_byte_index(&mut self,
i: ByteIndex,
character: char,
data: &GlyphData)

Adds a single glyph.

fn add_glyphs_for_byte_index(&mut self,
i: ByteIndex,
data_for_glyphs: &[GlyphData])

fn iter_glyphs_for_byte_range(&'a self,
range: &Range<ByteIndex>)
-> GlyphIterator<'a>

fn range_index_of_advance(&self,
range: &Range<ByteIndex>,
advance: Au,
extra_word_spacing: Au)
-> (usize, Au)

fn advance_for_byte_range(&self,
range: &Range<ByteIndex>,
extra_word_spacing: Au)
-> Au

fn advance_for_byte_range_slow_path(&self,
range: &Range<ByteIndex>,
extra_word_spacing: Au)
-> Au

fn advance_for_byte_range_simple_glyphs(&self,
range: &Range<ByteIndex>,
extra_word_spacing: Au)
-> Au

fn transmute_entry_buffer_to_u32_buffer(&self) -> &[u32]

Used for SIMD.

fn char_is_space(&self, i: ByteIndex) -> bool

fn space_count_in_range(&self, range: &Range<ByteIndex>) -> u32