Module gfx::text::text_run

Structs

CharacterSliceIterator

GlyphRun
A single series of glyphs within a text run.

LineIterator

NaturalWordSliceIterator

TextRun
A single "paragraph" of text in one font size and style.

TextRunSlice
A "slice" of a text run is a series of contiguous glyphs that all belong to the same glyph store. Line breaking strategies yield these.