Struct gfx::text::text_run::GlyphRun

pub struct GlyphRun {
    pub glyph_store: Arc<GlyphStore>,
    pub range: Range<ByteIndex>,
}

A single series of glyphs within a text run.

Fields

glyph_store

The glyphs.

range

The byte range of characters in the containing run.

Methods

impl GlyphRun

fn compare(&self, key: &ByteIndex) -> Ordering