Struct gfx::text::text_run::TextRun

pub struct TextRun {
    pub text: Arc<String>,
    pub font_template: Arc<FontTemplateData>,
    pub actual_pt_size: Au,
    pub font_metrics: FontMetrics,
    pub font_key: FontKey,
    pub glyphs: Arc<Vec<GlyphRun>>,
    pub bidi_level: u8,
    pub extra_word_spacing: Au,
}

A single "paragraph" of text in one font size and style.

Fields
text: Arc<String>

The UTF-8 string represented by this text run.

font_template: Arc<FontTemplateData> actual_pt_size: Au font_metrics: FontMetrics font_key: FontKey glyphs: Arc<Vec<GlyphRun>>

The glyph runs that make up this text run.

bidi_level: u8 extra_word_spacing: Au
Methods

impl<'a> TextRun

fn new(font: &mut Font,
text: String,
options: &ShapingOptions,
bidi_level: u8)
-> TextRun

fn break_and_shape(font: &mut Font,
text: &str,
options: &ShapingOptions)
-> Vec<GlyphRun>

fn ascent(&self) -> Au

fn descent(&self) -> Au

fn advance_for_range(&self, range: &Range<ByteIndex>) -> Au

fn metrics_for_range(&self, range: &Range<ByteIndex>) -> RunMetrics

fn metrics_for_slice(&self,
glyphs: &GlyphStore,
slice_range: &Range<ByteIndex>)
-> RunMetrics

fn min_width_for_range(&self, range: &Range<ByteIndex>) -> Au

fn minimum_splittable_inline_size(&self, range: &Range<ByteIndex>) -> Au

fn index_of_first_glyph_run_containing(&self, index: ByteIndex) -> Option<usize>

Returns the index of the first glyph run containing the given character index.

fn natural_word_slices_in_range(&'a self,
range: &Range<ByteIndex>)
-> NaturalWordSliceIterator<'a>

Returns an iterator that will iterate over all slices of glyphs that represent natural words in the given range.

fn natural_word_slices_in_visual_order(&'a self,
range: &Range<ByteIndex>)
-> NaturalWordSliceIterator<'a>

Returns an iterator that over natural word slices in visual order (left to right or right to left, depending on the bidirectional embedding level).

fn character_slices_in_range(&'a self,
range: &Range<ByteIndex>)
-> CharacterSliceIterator<'a>

Returns an iterator that will iterate over all slices of glyphs that represent individual characters in the given range.