Trait gfx::VertexFormat

pub trait VertexFormat {
    fn generate(Option<Self>, buffer: Handle<u32, BufferInfo>) -> Vec<Attribute>;
}

A trait implemented automatically for user vertex structure by `#[vertex_format] attribute

Required Methods

fn generate(Option<Self>, buffer: Handle<u32, BufferInfo>) -> Vec<Attribute>

Create the attributes for this type, using the given buffer.