Module gfx::traits

public re-exported traits

Traits

Device
A Device is responsible for submitting CommandBuffers to the GPU.

Factory
A Factory is responsible for creating and managing resources for the context it was created with.

FactoryExt
This trait is responsible for creating and managing graphics resources, much like the Factory trait in the gfx crate. Every Factory automatically implements FactoryExt.

Pod
A trait for plain-old-data types.