Crate gfx

An efficient, low-level, bindless graphics API for Rust. See the blog for explanations and annotated examples.

Modules

attrib
Vertex attribute types.

front
Frontend Rendering front-end

shade
Shaders Shader parameter handling.

state
Fixed-function hardware state.

tex
Texture creation and modification.

Structs

Attribute
Describes a single attribute of a vertex buffer, including its type, name, etc.

BufferHandle
Type-safe buffer handle

BufferInfo
An information block that is immutable and associated with each buffer

ClearData
How to clear a frame.

DrawState
An assembly of states that affect regular draw calls

Frame
A complete Frame, which is the result of rendering.

GlDevice
An OpenGL device with GLSL shaders

Mesh
Describes geometry to render.

ProgramInfo
Metadata about a program.

Renderer
Renderer front-end

ShaderSource
A type storing shader source for different graphics APIs and versions.

Enums

BufferUsage
A hint as to how this buffer will be used.

Plane
A single buffer that can be bound to a render target.

Slice
Description of a subset of Mesh data to render. We provide a primitive type in a slice because it is how we interpret mesh contents. For example, we can have a Point typed vertex slice to do shape blending, while still rendereing it as an indexed TriangleList.

UniformValue
A value that can be uploaded to the device as a uniform.

Traits

Blob
A trait that slice-like types implement.

Device
An interface for performing draw calls using a specific graphics API

DeviceHelper
Backend extension trait for convenience methods

VertexFormat
A trait implemented automatically for user vertex structure by `#[vertex_format] attribute

Type Definitions

Color

IndexCount

Layer

Level

ProgramHandle

RawBufferHandle

ShaderHandle

SurfaceHandle

TextureHandle

VertexCount