Module gfx::display_list

Servo heavily uses display lists, which are retained-mode lists of painting commands to perform. Using a list instead of painting elements in immediate mode allows transforms, hit testing, and invalidation to be performed using the same primitives as painting. It also allows Servo to aggressively cull invisible and out-of-bounds painting elements, to reduce overdraw. Finally, display lists allow tiles to be farmed out onto multiple CPUs and painted in parallel (although this benefit does not apply to GPU-based painting).

Display items describe relatively high-level drawing operations (for example, entire borders and shadows instead of lines and blur operations), to reduce the amount of allocation required. They are therefore not exactly analogous to constructs like Skia pictures, which consist of low-level drawing primitives.

Structs

BaseDisplayItem
Information common to all display items.

BorderDisplayItem
Paints a border.

BorderRadii
Information about the border radii.

BoxShadowDisplayItem
Paints a box shadow per CSS-BACKGROUNDS.

ClippingRegion
A clipping region for a display item. Currently, this can describe rectangles, rounded rectangles (for border-radius), or arbitrary intersections of the two. Arbitrary transforms are not supported because those are handled by the higher-level StackingContext abstraction.

ComplexClippingRegion
A complex clipping region. These don't as easily admit arbitrary intersection operations, so they're stored in a list over to the side. Currently a complex clipping region is just a rounded rectangle, but the CSS WGs will probably make us throw more stuff in here eventually.

DisplayItemMetadata
Metadata attached to each display item. This is useful for performing auxiliary threads with the display list involving hit testing: finding the originating DOM node and determining the cursor to use when the element is hovered over.

DisplayList

DisplayListTraversal

FnvHashMap
A FNV-based hash map. This is not serializable by serde by default, so we provide an implementation ourselves.

FnvHashMapVisitor
A visitor that produces a map.

GradientDisplayItem
Paints a gradient.

GradientStop

IframeDisplayItem
Paints an iframe.

ImageDisplayItem
Paints an image.

LayerInfo
LayerInfo is used to store PaintLayer metadata during DisplayList construction. It is also used for tracking LayerIds when creating layers to preserve ordering when layered DisplayItems should render underneath unlayered DisplayItems.

LayeredItem
Contains an item that should get its own layer during layer creation.

LineDisplayItem
Paints a line segment.

OpaqueNode
An opaque handle to a node, which, unlike UnsafeNode, cannot be transformed back into a non-opaque representation. The only safe operation that can be performed on this node is to compare it to another opaque handle or to another OpaqueNode.

SolidColorDisplayItem
Paints a solid color.

StackingContext
Represents one CSS stacking context, which may or may not have a hardware layer.

StackingContextOffsets

TextDisplayItem
Paints text.

WebGLDisplayItem

WebRenderImageInfo

Enums

BoxShadowClipMode
How a box shadow should be clipped.

DisplayItem
One drawing command in the list.

DisplayListSection
Display list sections that make up a stacking context. Each section here refers to the steps in CSS 2.1 Appendix E.

StackingContextType

TextOrientation

Constants

_IMPL_DESERIALIZE_FOR_BaseDisplayItem

_IMPL_DESERIALIZE_FOR_BorderDisplayItem

_IMPL_DESERIALIZE_FOR_BorderRadii

_IMPL_DESERIALIZE_FOR_BoxShadowClipMode

_IMPL_DESERIALIZE_FOR_BoxShadowDisplayItem

_IMPL_DESERIALIZE_FOR_ClippingRegion

_IMPL_DESERIALIZE_FOR_ComplexClippingRegion

_IMPL_DESERIALIZE_FOR_DisplayItem

_IMPL_DESERIALIZE_FOR_DisplayItemMetadata

_IMPL_DESERIALIZE_FOR_DisplayList

_IMPL_DESERIALIZE_FOR_DisplayListSection

_IMPL_DESERIALIZE_FOR_GradientDisplayItem

_IMPL_DESERIALIZE_FOR_IframeDisplayItem

_IMPL_DESERIALIZE_FOR_ImageDisplayItem

_IMPL_DESERIALIZE_FOR_LayerInfo

_IMPL_DESERIALIZE_FOR_LayeredItem

_IMPL_DESERIALIZE_FOR_LineDisplayItem

_IMPL_DESERIALIZE_FOR_SolidColorDisplayItem

_IMPL_DESERIALIZE_FOR_StackingContext

_IMPL_DESERIALIZE_FOR_StackingContextOffsets

_IMPL_DESERIALIZE_FOR_StackingContextType

_IMPL_DESERIALIZE_FOR_TextDisplayItem

_IMPL_DESERIALIZE_FOR_TextOrientation

_IMPL_DESERIALIZE_FOR_WebGLDisplayItem

_IMPL_DESERIALIZE_FOR_WebRenderImageInfo

_IMPL_SERIALIZE_FOR_BaseDisplayItem

_IMPL_SERIALIZE_FOR_BorderDisplayItem

_IMPL_SERIALIZE_FOR_BorderRadii

_IMPL_SERIALIZE_FOR_BoxShadowClipMode

_IMPL_SERIALIZE_FOR_BoxShadowDisplayItem

_IMPL_SERIALIZE_FOR_ClippingRegion

_IMPL_SERIALIZE_FOR_ComplexClippingRegion

_IMPL_SERIALIZE_FOR_DisplayItem

_IMPL_SERIALIZE_FOR_DisplayItemMetadata

_IMPL_SERIALIZE_FOR_DisplayList

_IMPL_SERIALIZE_FOR_DisplayListSection

_IMPL_SERIALIZE_FOR_GradientDisplayItem

_IMPL_SERIALIZE_FOR_IframeDisplayItem

_IMPL_SERIALIZE_FOR_ImageDisplayItem

_IMPL_SERIALIZE_FOR_LayerInfo

_IMPL_SERIALIZE_FOR_LayeredItem

_IMPL_SERIALIZE_FOR_LineDisplayItem

_IMPL_SERIALIZE_FOR_SolidColorDisplayItem

_IMPL_SERIALIZE_FOR_StackingContext

_IMPL_SERIALIZE_FOR_StackingContextOffsets

_IMPL_SERIALIZE_FOR_StackingContextType

_IMPL_SERIALIZE_FOR_TextDisplayItem

_IMPL_SERIALIZE_FOR_TextOrientation

_IMPL_SERIALIZE_FOR_WebGLDisplayItem

_IMPL_SERIALIZE_FOR_WebRenderImageInfo

Statics

BLUR_INFLATION_FACTOR
The factor that we multiply the blur radius by in order to inflate the boundaries of display items that involve a blur. This ensures that the display item boundaries include all the ink.

Functions

transformed_tile_rect

Type Definitions

ScrollOffsetMap
The type of the scroll offset list. This is only populated if WebRender is in use.

Implementations