Struct gfx::display_list::StackingContext

pub struct StackingContext {
    pub id: StackingContextId,
    pub context_type: StackingContextType,
    pub bounds: Rect<Au>,
    pub overflow: Rect<Au>,
    pub z_index: i32,
    pub filters: T,
    pub blend_mode: T,
    pub transform: Matrix4D<f32>,
    pub perspective: Matrix4D<f32>,
    pub establishes_3d_context: bool,
    pub scrolls_overflow_area: bool,
    pub layer_info: Option<LayerInfo>,
    pub children: Vec<Box<StackingContext>>,
}

Represents one CSS stacking context, which may or may not have a hardware layer.

Fields

id

The ID of this StackingContext for uniquely identifying it.

context_type

The type of this StackingContext. Used for collecting and sorting.

bounds

The position and size of this stacking context.

overflow

The overflow rect for this stacking context in its coordinate system.

z_index

The z-index for this stacking context.

filters

CSS filters to be applied to this stacking context (including opacity).

blend_mode

The blend mode with which this stacking context blends with its backdrop.

transform

A transform to be applied to this stacking context.

perspective

The perspective matrix to be applied to children.

establishes_3d_context

Whether this stacking context creates a new 3d rendering context.

scrolls_overflow_area

Whether this stacking context scrolls its overflow area.

layer_info

The layer info for this stacking context, if there is any.

children

Children of this StackingContext.

Methods

impl StackingContext

fn new(id: StackingContextId, context_type: StackingContextType, bounds: &Rect<Au>, overflow: &Rect<Au>, z_index: i32, filters: T, blend_mode: T, transform: Matrix4D<f32>, perspective: Matrix4D<f32>, establishes_3d_context: bool, scrolls_overflow_area: bool, layer_info: Option<LayerInfo>) -> StackingContext

Creates a new stacking context.

fn hit_test<'a>(&self, traversal: &mut DisplayListTraversal<'a>, point: Point2D<Au>, result: &mut Vec<DisplayItemMetadata>)

fn print_with_tree(&self, print_tree: &mut PrintTree)

fn intersects_rect_in_parent_context(&self, rect: Option<Rect<Au>>) -> bool