Struct gfx::display_list::StackingContext

pub struct StackingContext {
    pub display_list: Box<DisplayList>,
    pub layer: Option<Arc<PaintLayer>>,
    pub bounds: Rect<Au>,
    pub overflow: Rect<Au>,
    pub z_index: i32,
    pub filters: T,
    pub blend_mode: T,
    pub transform: Matrix2D<AzFloat>,
}

Represents one CSS stacking context, which may or may not have a hardware layer.

Fields

display_list

The display items that make up this stacking context.

layer

The layer for this stacking context, if there is one.

bounds

The position and size of this stacking context.

overflow

The overflow rect for this stacking context in its coordinate system.

z_index

The z-index for this stacking context.

filters

CSS filters to be applied to this stacking context (including opacity).

blend_mode

The blend mode with which this stacking context blends with its backdrop.

transform

A transform to be applied to this stacking context.

TODO(pcwalton): 3D transforms.

Methods

impl StackingContext

fn new(display_list: Box<DisplayList>, bounds: &Rect<Au>, overflow: &Rect<Au>, z_index: i32, transform: &Matrix2D<AzFloat>, filters: T, blend_mode: T, layer: Option<Arc<PaintLayer>>) -> StackingContext

Creates a new stacking context.

fn optimize_and_draw_into_context(&self, paint_context: &mut PaintContext, tile_bounds: &Rect<AzFloat>, transform: &Matrix2D<AzFloat>, clip_rect: Option<&Rect<Au>>)

Draws the stacking context in the proper order according to the steps in CSS 2.1  E.2.

fn hit_test(&self, point: Point2D<Au>, result: &mut Vec<DisplayItemMetadata>, topmost_only: bool)

Places all nodes containing the point of interest into result, topmost first. Respects the pointer-events CSS property If topmost_only is true, stops after placing one node into the list. result must be empty upon entry to this function.