Module gfx::font

Macros

ot_tag

Structs

Font

FontDescriptor

FontDescriptor describes the parameters of a Font. It represents rendering a given font template at a particular size, with a particular font-variant-caps applied, etc. This contrasts with FontTemplateDescriptor in that the latter represents only the parameters inherent in the font data (weight, stretch, etc.).

FontFamilyDescriptor

The font family parameters for font selection.

FontGroup

A FontGroup is a prioritised list of fonts for a given set of font styles. It is used by TextRun to decide which font to render a character with. If none of the fonts listed in the styles are suitable, a fallback font may be used.

FontGroupFamily

A FontGroupFamily is a single font family in a FontGroup. It corresponds to one of the families listed in the font-family CSS property. The corresponding font data is lazy-loaded, only if actually needed.

FontMetrics

RunMetrics

ShapeCacheEntry

An entry in the shape cache.

ShapingFlags

ShapingOptions

Various options that control text shaping.

Enums

FontFamilyName

A font family name used in font selection.

FontSearchScope

The scope within which we will look for a font.

Constants

GPOS

GSUB

KERN

LAST_RESORT_GLYPH_ADVANCE

Statics

TEXT_SHAPING_PERFORMANCE_COUNTER

Traits

FontHandleMethods

FontTableMethods

FontTableTagConversions

Functions

fallback_font_families

get_and_reset_text_shaping_performance_counter

Type Definitions

FontRef

FontTableTag

FractionalPixel