Module gfx::font_cache_thread

Structs

FontCache
The font cache thread itself. It maintains a list of reference counted font templates that are currently in use.

FontCacheThread
The public interface to the font cache thread, used exclusively by the per-thread/thread FontContext structures.

FontTemplateInfo

FontTemplates
A list of font templates that make up a given font family.

LowercaseString

Enums

Command
Commands that the FontContext sends to the font cache thread.

Reply
Reply messages sent from the font cache thread to the FontContext caller.

Constants

_IMPL_DESERIALIZE_FOR_Command

_IMPL_DESERIALIZE_FOR_FontCacheThread

_IMPL_DESERIALIZE_FOR_FontTemplateInfo

_IMPL_DESERIALIZE_FOR_LowercaseString

_IMPL_DESERIALIZE_FOR_Reply

_IMPL_SERIALIZE_FOR_Command

_IMPL_SERIALIZE_FOR_FontCacheThread

_IMPL_SERIALIZE_FOR_FontTemplateInfo

_IMPL_SERIALIZE_FOR_LowercaseString

_IMPL_SERIALIZE_FOR_Reply

Functions

is_supported_font_type

populate_generic_fonts