Module gfx::font_context

Structs

FallbackFontCacheEntry

FontContext
The FontContext represents the per-thread/thread state necessary for working with fonts. It is the public API used by the layout and paint code. It talks directly to the font cache thread where required.

LayoutFontCacheEntry

LayoutFontGroupCacheKey

PaintFontCacheEntry
A cached azure font (per paint thread) that can be shared by multiple text runs.

Statics

FONT_CACHE_EPOCH
An epoch for the font context cache. The cache is flushed if the current epoch does not match this one.

SMALL_CAPS_SCALE_FACTOR

Functions

create_scaled_font

invalidate_font_caches