Struct gfx::font_context::FontContext

pub struct FontContext {
    platform_handle: FontContextHandle,
    font_cache_thread: FontCacheThread,
    layout_font_cache: Vec<LayoutFontCacheEntry>,
    fallback_font_cache: Vec<FallbackFontCacheEntry>,
    paint_font_cache: Vec<PaintFontCacheEntry>,
    layout_font_group_cache: HashMap<LayoutFontGroupCacheKey, Rc<FontGroup>, BuildHasherDefault<FnvHasher>>,
    epoch: usize,
}

The FontContext represents the per-thread/thread state necessary for working with fonts. It is the public API used by the layout and paint code. It talks directly to the font cache thread where required.

Fields

platform_handle

font_cache_thread

layout_font_cache

TODO: See bug https://github.com/servo/servo/issues/3300.

fallback_font_cache

paint_font_cache

Strong reference as the paint FontContext is (for now) recycled per frame. TODO: Make this weak when incremental redraw is done.

layout_font_group_cache

epoch

Methods

impl FontContext

fn new(font_cache_thread: FontCacheThread) -> FontContext

fn create_layout_font(&self, template: Arc<FontTemplateData>, descriptor: FontTemplateDescriptor, pt_size: Au, variant: T, font_key: Option<FontKey>) -> Result<Font, ()>

Create a font for use in layout calculations.

fn expire_font_caches_if_necessary(&mut self)

fn layout_font_group_for_style(&mut self, style: Arc<ServoFont>) -> Rc<FontGroup>

Create a group of fonts for use in layout calculations. May return a cached font if this font instance has already been used by this context.

fn paint_font_from_template(&mut self, template: &Arc<FontTemplateData>, pt_size: Au) -> Rc<RefCell<ScaledFont>>

Create a paint font for use with azure. May return a cached reference if already used by this font context.

fn font_cache_thread(&self) -> FontCacheThread

Returns a reference to the font cache thread.