Struct gfx::font_context::FontContext

pub struct FontContext {
    // some fields omitted
}

The FontContext represents the per-thread/task state necessary for working with fonts. It is the public API used by the layout and paint code. It talks directly to the font cache task where required.

Methods

impl FontContext

fn new(font_cache_task: FontCacheTask) -> FontContext

fn get_layout_font_group_for_style(&mut self, style: Arc<SpecifiedFontStyle>) -> Rc<FontGroup>

Create a group of fonts for use in layout calculations. May return a cached font if this font instance has already been used by this context.

fn get_paint_font_from_template(&mut self, template: &Arc<FontTemplateData>, pt_size: Au) -> Rc<RefCell<ScaledFont>>

Create a paint font for use with azure. May return a cached reference if already used by this font context.

fn font_cache_task(&self) -> FontCacheTask

Returns a reference to the font cache task.