Struct gfx::font_template::FontTemplate

pub struct FontTemplate {
    identifier: Atom,
    descriptor: Option<FontTemplateDescriptor>,
    weak_ref: Option<Weak<FontTemplateData>>,
    strong_ref: Option<Arc<FontTemplateData>>,
    is_valid: bool,
}

This describes all the information needed to create font instance handles. It contains a unique FontTemplateData structure that is platform specific.

Fields

identifier

descriptor

weak_ref

strong_ref

is_valid

Methods

impl FontTemplate

Holds all of the template information for a font that is common, regardless of the number of instances of this font handle per thread.

fn new(identifier: Atom, maybe_bytes: Option<Vec<u8>>) -> FontTemplate

fn identifier(&self) -> &Atom

fn data_for_descriptor(&mut self, fctx: &FontContextHandle, requested_desc: &FontTemplateDescriptor) -> Option<Arc<FontTemplateData>>

Get the data for creating a font if it matches a given descriptor.

fn data_for_approximate_descriptor(&mut self, font_context: &FontContextHandle, requested_descriptor: &FontTemplateDescriptor) -> Option<(Arc<FontTemplateData>, u32)>

Returns the font data along with the distance between this font's descriptor and the given descriptor, if the font can be loaded.

fn instantiate(&mut self, font_context: &FontContextHandle) -> Result<(), ()>

fn get(&mut self) -> Option<Arc<FontTemplateData>>

Get the data for creating a font.

fn data(&mut self) -> Arc<FontTemplateData>

Get the font template data. If any strong references still exist, it will return a clone, otherwise it will load the font data and store a weak reference to it internally.