Struct gfx::font_template::FontTemplate

pub struct FontTemplate {
    // some fields omitted
}

This describes all the information needed to create font instance handles. It contains a unique FontTemplateData structure that is platform specific.

Methods

impl FontTemplate

Holds all of the template information for a font that is common, regardless of the number of instances of this font handle per thread.

fn new(identifier: &str, maybe_bytes: Option<Vec<u8>>) -> FontTemplate

fn identifier<'a>(&'a self) -> &'a str

fn get_if_matches(&mut self, fctx: &FontContextHandle, requested_desc: &FontTemplateDescriptor) -> Option<Arc<FontTemplateData>>

Get the data for creating a font if it matches a given descriptor.

fn get(&mut self) -> Option<Arc<FontTemplateData>>

Get the data for creating a font.

fn get_data(&mut self) -> Arc<FontTemplateData>

Get the font template data. If any strong references still exist, it will return a clone, otherwise it will load the font data and store a weak reference to it internally.