Struct gfx::font_template::FontTemplateDescriptor

pub struct FontTemplateDescriptor {
    pub weight: T,
    pub stretch: T,
    pub italic: bool,
}

Describes how to select a font from a given family. This is very basic at the moment and needs to be expanded or refactored when we support more of the font styling parameters.

NB: If you change this, you will need to update style::properties::compute_font_hash().

Fields

weight

stretch

italic

Methods

impl FontTemplateDescriptor

fn new(weight: T, stretch: T, italic: bool) -> FontTemplateDescriptor

fn distance_from(&self, other: &FontTemplateDescriptor) -> u32

Returns a score indicating how far apart visually the two font descriptors are. This is used for fuzzy font selection.

The smaller the score, the better the fonts match. 0 indicates an exact match. This must be commutative (distance(A, B) == distance(B, A)).