Struct gfx::handle::Manager

pub struct Manager<R> where
    R: Resources,  { /* private fields */ }

Stores reference-counted resources used in a command buffer. Seals actual resource names behind the interface, automatically referencing them both by the Factory on resource creation and the Renderer during CommandBuffer population.

Implementations

impl<R> Manager<R> where
R: Resources,

pub fn new() -> Manager<R>

Create a new handle manager

pub fn clear(&mut self)

Clear all references

pub fn extend(&mut self, other: &Manager<R>)

Extend with all references of another handle manager

pub fn count(&self) -> usize

Count the total number of referenced resources

pub fn ref_buffer(
&mut self,
handle: &'a RawBuffer<R>
) -> &'a <R as Resources>::Buffer

Reference a buffer

pub fn ref_shader(
&mut self,
handle: &'a Shader<R>
) -> &'a <R as Resources>::Shader

Reference a shader

pub fn ref_program(
&mut self,
handle: &'a Program<R>
) -> &'a <R as Resources>::Program

Reference a program

pub fn ref_pso(
&mut self,
handle: &'a RawPipelineState<R>
) -> (&'a <R as Resources>::PipelineStateObject, &'a <R as Resources>::Program)

Reference a pipeline state object

pub fn ref_texture(
&mut self,
handle: &'a RawTexture<R>
) -> &'a <R as Resources>::Texture

Reference a texture

pub fn ref_srv(
&mut self,
handle: &'a RawShaderResourceView<R>
) -> &'a <R as Resources>::ShaderResourceView

Reference a shader resource view

pub fn ref_uav(
&mut self,
handle: &'a RawUnorderedAccessView<R>
) -> &'a <R as Resources>::UnorderedAccessView

Reference an unordered access view

pub fn ref_rtv(
&mut self,
handle: &'a RawRenderTargetView<R>
) -> &'a <R as Resources>::RenderTargetView

Reference an RTV

pub fn ref_dsv(
&mut self,
handle: &'a RawDepthStencilView<R>
) -> &'a <R as Resources>::DepthStencilView

Reference a DSV

pub fn ref_sampler(
&mut self,
handle: &'a Sampler<R>
) -> &'a <R as Resources>::Sampler

Reference a sampler

pub fn ref_fence(&mut self, fence: &'a Fence<R>) -> &'a <R as Resources>::Fence

Reference a fence