Struct gfx::handle::Manager

pub struct Manager<R> where R: Resources {
    // some fields omitted
}

Stores reference-counted resources used in a command buffer. Seals actual resource names behind the interface, automatically referencing them both by the Factory on resource creation and the Renderer during CommandBuffer population.

Methods

impl<R> Manager<R> where R: Resources

fn new() -> Manager<R>

Create a new handle manager

fn clear(&mut self)

Clear all references

fn extend(&mut self, other: &Manager<R>)

Extend with all references of another handle manager

fn count(&self) -> usize

Count the total number of referenced resources

fn ref_buffer(&mut self, handle: &'a RawBuffer<R>) -> &'a R::Buffer

Reference a buffer

fn ref_shader(&mut self, handle: &'a Shader<R>) -> &'a R::Shader

Reference a shader

fn ref_program(&mut self, handle: &'a Program<R>) -> &'a R::Program

Reference a program

fn ref_pso(&mut self, handle: &'a RawPipelineState<R>) -> (&'a R::PipelineStateObject, &'a R::Program)

Reference a pipeline state object

fn ref_texture(&mut self, handle: &'a RawTexture<R>) -> &'a R::Texture

Reference a texture

fn ref_srv(&mut self, handle: &'a RawShaderResourceView<R>) -> &'a R::ShaderResourceView

Reference a shader resource view

fn ref_uav(&mut self, handle: &'a RawUnorderedAccessView<R>) -> &'a R::UnorderedAccessView

Reference an unordered access view

fn ref_rtv(&mut self, handle: &'a RawRenderTargetView<R>) -> &'a R::RenderTargetView

Reference an RTV

fn ref_dsv(&mut self, handle: &'a RawDepthStencilView<R>) -> &'a R::DepthStencilView

Reference a DSV

fn ref_sampler(&mut self, handle: &'a Sampler<R>) -> &'a R::Sampler

Reference a sampler

fn ref_fence(&mut self, fence: &'a Fence<R>) -> &'a R::Fence

Reference a fence