Struct gfx::platform::freetype::font_template::FontTemplateData

pub struct FontTemplateData {
    pub bytes: Vec<u8>,
    pub identifier: String,
}

Platform specific font representation for Linux. The identifier is an absolute path, and the bytes field is the loaded data that can be passed to freetype and azure directly.

Fields

bytes

identifier

Methods

impl FontTemplateData

fn new(identifier: &str, font_data: Option<Vec<u8>>) -> FontTemplateData