Module gfx::pso::buffer

Buffer components for a PSO.

Structs

ConstantBuffer

Constant buffer component.

Element

A struct element descriptor.

Global

Global (uniform) constant component. Describes a free-standing value passed into the shader, which is not enclosed into any constant buffer.

RawConstantBuffer

Raw constant buffer component.

RawGlobal

Raw global (uniform) constant component. Describes a free-standing value passed into the shader, which is not enclosed in any constant buffer.

RawVertexBuffer

Raw vertex/instance buffer component. Can be used when the formats of vertex attributes are not known at compile time.

VertexBufferCommon

Service struct to simplify the implementations of VertexBuffer and InstanceBuffer.

Enums

Instanced

Helper phantom type for per-instance attributes.

NonInstanced

Helper phantom type for per-vertex attributes.

Traits

Structure

A trait to be implemented by any struct having the layout described in the graphics API, like a vertex buffer.

ToInstanceRate

Helper trait for VertexBufferCommon to support variable instance rate.

Type Definitions

BufferIndex

Index of a vertex buffer.

ElemOffset

Offset of an attribute from the start of the buffer, in bytes

ElemStride

Offset between attribute values, in bytes

InstanceBuffer

Instance buffer component. Same as the vertex buffer but advances per instance.

InstanceRate

The number of instances between each subsequent attribute value

VertexBuffer

Vertex buffer component. Advanced per vertex.