Struct gfx::pso::RawDataSet

pub struct RawDataSet<R: Resources> {
    pub vertex_buffers: VertexBufferSet<R>,
    pub constant_buffers: Vec<ConstantBufferParam<R>>,
    pub global_constants: Vec<(Location, UniformValue)>,
    pub resource_views: Vec<ResourceViewParam<R>>,
    pub unordered_views: Vec<UnorderedViewParam<R>>,
    pub samplers: Vec<SamplerParam<R>>,
    pub pixel_targets: PixelTargetSet<R>,
    pub ref_values: RefValues,
    pub scissor: Rect,
}

A complete set of raw data that needs to be specified at run-time whenever we draw something with a PSO. This is what "data" struct gets transformed into when we call encoder.draw(...) with it. It doesn't have any typing information, since PSO knows what format and layout to expect from each resource.

Fields
vertex_buffers: VertexBufferSet<R> constant_buffers: Vec<ConstantBufferParam<R>> global_constants: Vec<(Location, UniformValue)> resource_views: Vec<ResourceViewParam<R>> unordered_views: Vec<UnorderedViewParam<R>> samplers: Vec<SamplerParam<R>> pixel_targets: PixelTargetSet<R> ref_values: RefValues scissor: Rect
Methods

impl<R: Resources> RawDataSet<R>

pub fn new() -> RawDataSet<R>

Create an empty data set.

pub fn clear(&mut self)

Clear all contained data.