Trait gfx::pso::DataBind

pub trait DataBind<R: Resources> {
    type Data;

    fn bind_to(
        &self, 
        _: &mut RawDataSet<R>, 
        _: &Self::Data, 
        _: &mut Manager<R>, 
        _: &mut AccessInfo<R>
    );
}

The “bind” logic portion of the PSO component. Defines how the user data translates into the raw data set.

Required Associated Types

type Data

The associated “data” type - a member of the PSO “data” struct.

Required Methods

fn bind_to(
&self,
_: &mut RawDataSet<R>,
_: &Self::Data,
_: &mut Manager<R>,
_: &mut AccessInfo<R>
)

Dump the given data into the raw data set.