Struct gfx::render::Renderer

pub struct Renderer<R: Resources, C: CommandBuffer<R>> {
    // some fields omitted
}

Renderer front-end

Methods

impl<R: Resources, C: CommandBuffer<R>> Renderer<R, C>

fn reset(&mut self)

Reset all commands for the command buffer re-usal.

fn as_buffer(&self) -> (&C, &DataBuffer, &Manager<R>)

Get command and data buffers to be submitted to the device.

fn clone_empty(&self) -> Renderer<R, C>

Clone the renderer shared data but ignore the commands.

fn clear<O: Output<R>>(&mut self, data: ClearData, mask: Mask, output: &O)

Clear the output with given ClearData.

fn draw<B: Batch<R> + ?Sized, O: Output<R>>(&mut self, batch: &B, instances: InstanceOption, output: &O) -> Result<(), DrawError<Error>>

Draw a 'batch' with all known parameters specified, internal use only.

fn blit<I: Output<R>, O: Output<R>>(&mut self, source: &I, source_rect: Rect, destination: &O, dest_rect: Rect, mirror: Mirror, mask: Mask) -> Result<(), BlitError>

Blit one frame onto another.

fn update_buffer<T: Copy>(&mut self, buf: &RawBuffer<R>, data: &[T], offset_elements: usize) -> Result<(), UpdateError<usize>>

Update a buffer with a slice of data.

fn update_block<U, T: Copy>(&mut self, buf: &Buffer<R, U>, data: &T) -> Result<(), UpdateError<usize>>

Update a buffer with a data struct.

fn update_texture<T: Copy>(&mut self, tex: &Texture<R>, img: ImageInfo, data: &[T]) -> Result<(), UpdateError<[Size; 3]>>

Update the contents of a texture.