Struct gfx::Encoder

pub struct Encoder<R: Resources, C: CommandBuffer<R>> {
    // some fields omitted
}

Graphics commands encoder.

Methods

impl<R: Resources, C: CommandBuffer<R>> Encoder<R, C>

fn flush<D>(&mut self, device: &mut D) where D: Device<Resources=R, CommandBuffer=C>

Flush the encoded commands onto the device, and clean.

fn clone_empty(&self) -> Encoder<R, C>

Clone the renderer shared data but ignore the commands.

fn update_buffer<T: Copy>(&mut self, buf: &Buffer<R, T>, data: &[T], offset_elements: usize) -> Result<(), UpdateError<usize>>

Update a buffer with a slice of data.

fn update_constant_buffer<T: Copy>(&mut self, buf: &Buffer<R, T>, data: &T)

Update a buffer with a single structure.

fn update_texture<S, T>(&mut self, tex: &Texture<R, T::Surface>, face: Option<CubeFace>, img: NewImageInfo, data: &[S::DataType]) -> Result<(), UpdateError<[Size; 3]>> where S: SurfaceTyped, S::DataType: Copy, T: Formatted<Surface=S>

Update the contents of a texture.

fn clear<T: RenderFormat>(&mut self, view: &RenderTargetView<R, T>, value: T::View) where T::View: Into<ClearColor>

Clear a target view with a specified value.

fn clear_depth<T: DepthFormat>(&mut self, view: &DepthStencilView<R, T>, depth: Depth)

Clear a depth view with a specified value.

fn clear_stencil<T: StencilFormat>(&mut self, view: &DepthStencilView<R, T>, stencil: Stencil)

Clear a stencil view with a specified value.

fn draw<D: PipelineData<R>>(&mut self, slice: &Slice<R>, pipeline: &PipelineState<R, D::Meta>, user_data: &D)

Draw a mesh slice using a typed pipeline state object (PSO).