Struct gfx::ProgramInfo

pub struct ProgramInfo {
    pub vertex_attributes: Vec<AttributeVar>,
    pub globals: Vec<ConstVar>,
    pub constant_buffers: Vec<ConstantBufferVar>,
    pub textures: Vec<TextureVar>,
    pub unordereds: Vec<UnorderedVar>,
    pub samplers: Vec<SamplerVar>,
    pub outputs: Vec<OutputVar>,
    pub knows_outputs: bool,
}

Metadata about a program.

Fields

vertex_attributes

Attributes in the program

globals

Global constants in the program

constant_buffers

Constant buffers in the program

textures

Textures in the program

unordereds

Unordered access resources in the program

samplers

Samplers in the program

outputs

Output targets in the program

knows_outputs

A hacky flag to make sure the clients know we are unable to actually get the output variable info