Module gfx::text::glyph

Structs

CharIndex
An index that refers to a character in a text run. This could point to the middle of a glyph.

GlyphData

GlyphIterator
An iterator over the glyphs in a character range in a GlyphStore.

GlyphStore
Stores the glyph data belonging to a text run.

Enums

BreakType

GlyphInfo

Type Definitions

GlyphId
The id of a particular glyph within a font