Struct gfx::text::glyph::GlyphStore

pub struct GlyphStore {
    // some fields omitted
}

Stores the glyph data belonging to a text run.

Simple glyphs are stored inline in the entry_buffer, detailed glyphs are stored as pointers into the detail_store.

+- GlyphStore --------------------------------+
|               +---+---+---+---+---+---+---+ |
| entry_buffer: |   | s |   | s |   | s | s | |  d = detailed
|               +-|-+---+-|-+---+-|-+---+---+ |  s = simple
|                 |       |       |           |
|                 |   +---+-------+           |
|                 |   |                       |
|               +-V-+-V-+                     |
| detail_store: | d | d |                     |
|               +---+---+                     |
+---------------------------------------------+

Methods

impl<'a> GlyphStore

fn new(length: isize, is_whitespace: bool) -> GlyphStore

fn char_len(&self) -> CharIndex

fn is_whitespace(&self) -> bool

fn finalize_changes(&mut self)

fn add_glyph_for_char_index(&mut self, i: CharIndex, character: Option<char>, data: &GlyphData)

Adds a single glyph. If character is present, this represents a single character; otherwise, this glyph represents multiple characters.

fn add_glyphs_for_char_index(&mut self, i: CharIndex, data_for_glyphs: &[GlyphData])

fn add_nonglyph_for_char_index(&mut self, i: CharIndex, cluster_start: bool, ligature_start: bool)

fn iter_glyphs_for_char_index(&'a self, i: CharIndex) -> GlyphIterator<'a>

fn iter_glyphs_for_char_range(&'a self, rang: &Range<CharIndex>) -> GlyphIterator<'a>

fn advance_for_char_range(&self, rang: &Range<CharIndex>) -> Au

fn char_is_space(&self, i: CharIndex) -> bool

fn char_is_tab(&self, i: CharIndex) -> bool

fn char_is_newline(&self, i: CharIndex) -> bool

fn is_ligature_start(&self, i: CharIndex) -> bool

fn is_cluster_start(&self, i: CharIndex) -> bool

fn can_break_before(&self, i: CharIndex) -> BreakType

fn set_char_is_space(&mut self, i: CharIndex)

fn set_char_is_tab(&mut self, i: CharIndex)

fn set_char_is_newline(&mut self, i: CharIndex)

fn set_can_break_before(&mut self, i: CharIndex, t: BreakType)

fn space_count_in_range(&self, range: &Range<CharIndex>) -> u32

fn distribute_extra_space_in_range(&mut self, range: &Range<CharIndex>, space: f64)