Module gfx::text::text_run

Structs

CharacterSliceIterator

GlyphRun
A single series of glyphs within a text run.

NaturalWordSliceIterator

TextRun
A single "paragraph" of text in one font size and style.

TextRunSlice
A "slice" of a text run is a series of contiguous glyphs that all belong to the same glyph store. Line breaking strategies yield these.

Constants

_IMPL_DESERIALIZE_FOR_GlyphRun

_IMPL_DESERIALIZE_FOR_TextRun

_IMPL_SERIALIZE_FOR_GlyphRun

_IMPL_SERIALIZE_FOR_TextRun

Statics

INDEX_OF_FIRST_GLYPH_RUN_CACHE