Struct gfx::text::text_run::TextRun

pub struct TextRun {
    pub text: Arc<String>,
    pub font_template: Arc<FontTemplateData>,
    pub actual_pt_size: Au,
    pub font_metrics: FontMetrics,
    pub font_key: FontInstanceKey,
    pub glyphs: Arc<Vec<GlyphRun>>,
    pub bidi_level: Level,
    pub extra_word_spacing: Au,
}

A single "paragraph" of text in one font size and style.

Fields
text: Arc<String>

The UTF-8 string represented by this text run.

font_template: Arc<FontTemplateData>actual_pt_size: Aufont_metrics: FontMetricsfont_key: FontInstanceKeyglyphs: Arc<Vec<GlyphRun>>

The glyph runs that make up this text run.

bidi_level: Levelextra_word_spacing: Au
Implementations

impl<'a> TextRun

pub fn new(
font: &mut Font,
text: String,
options: &ShapingOptions,
bidi_level: Level,
breaker: &mut Option<LineBreakLeafIter>
) -> (TextRun, bool)

Constructs a new text run. Also returns if there is a line break at the beginning

pub fn break_and_shape(
font: &mut Font,
text: &str,
options: &ShapingOptions,
breaker: &mut Option<LineBreakLeafIter>
) -> (Vec<GlyphRun>, bool)

pub fn ascent(&self) -> Au

pub fn advance_for_range(&self, range: &Range<ByteIndex>) -> Au

pub fn metrics_for_range(&self, range: &Range<ByteIndex>) -> RunMetrics

pub fn metrics_for_slice(
&self,
glyphs: &GlyphStore,
slice_range: &Range<ByteIndex>
) -> RunMetrics

pub fn min_width_for_range(&self, range: &Range<ByteIndex>) -> Au

pub fn minimum_splittable_inline_size(&self, range: &Range<ByteIndex>) -> Au

fn index_of_first_glyph_run_containing(&self, index: ByteIndex) -> Option<usize>

Returns the index of the first glyph run containing the given character index.

pub fn on_glyph_run_boundary(&self, index: ByteIndex) -> bool

pub fn range_index_of_advance(
&self,
range: &Range<ByteIndex>,
advance: Au
) -> usize

Returns the index in the range of the first glyph advancing over given advance

pub fn natural_word_slices_in_range(
&'a self,
range: &Range<ByteIndex>
) -> NaturalWordSliceIterator<'a>ⓘ

Important traits for NaturalWordSliceIterator<'a>
impl<'a> Iterator for NaturalWordSliceIterator<'a> type Item = TextRunSlice<'a>;

Returns an iterator that will iterate over all slices of glyphs that represent natural words in the given range.

pub fn natural_word_slices_in_visual_order(
&'a self,
range: &Range<ByteIndex>
) -> NaturalWordSliceIterator<'a>ⓘ

Important traits for NaturalWordSliceIterator<'a>
impl<'a> Iterator for NaturalWordSliceIterator<'a> type Item = TextRunSlice<'a>;

Returns an iterator that over natural word slices in visual order (left to right or right to left, depending on the bidirectional embedding level).

pub fn character_slices_in_range(
&'a self,
range: &Range<ByteIndex>
) -> CharacterSliceIterator<'a>ⓘ

Important traits for CharacterSliceIterator<'a>
impl<'a> Iterator for CharacterSliceIterator<'a> type Item = TextRunSlice<'a>;

Returns an iterator that will iterate over all slices of glyphs that represent individual characters in the given range.