Struct gfx::text::text_run::TextRun

pub struct TextRun {
    pub text: Arc<String>,
    pub font_template: Arc<FontTemplateData>,
    pub actual_pt_size: Au,
    pub font_metrics: FontMetrics,
    pub glyphs: Arc<Vec<GlyphRun>>,
}

A single "paragraph" of text in one font size and style.

Fields

text

font_template

actual_pt_size

font_metrics

glyphs

The glyph runs that make up this text run.

Methods

impl<'a> TextRun

fn new(font: &mut Font, text: String, options: &ShapingOptions) -> TextRun

fn break_and_shape(font: &mut Font, text: &str, options: &ShapingOptions) -> Vec<GlyphRun>

fn char_len(&self) -> CharIndex

fn glyphs(&'a self) -> &'a Vec<GlyphRun>

fn range_is_trimmable_whitespace(&self, range: &Range<CharIndex>) -> bool

fn ascent(&self) -> Au

fn descent(&self) -> Au

fn advance_for_range(&self, range: &Range<CharIndex>) -> Au

fn metrics_for_range(&self, range: &Range<CharIndex>) -> RunMetrics

fn metrics_for_slice(&self, glyphs: &GlyphStore, slice_range: &Range<CharIndex>) -> RunMetrics

fn min_width_for_range(&self, range: &Range<CharIndex>) -> Au

fn first_glyph_run_containing(&'a self, index: CharIndex) -> Option<&'a GlyphRun>

Returns the first glyph run containing the given character index.

fn natural_word_slices_in_range(&'a self, range: &Range<CharIndex>) -> NaturalWordSliceIterator<'a>

Returns an iterator that will iterate over all slices of glyphs that represent natural words in the given range.

fn character_slices_in_range(&'a self, range: &Range<CharIndex>) -> CharacterSliceIterator<'a>

Returns an iterator that will iterate over all slices of glyphs that represent individual characters in the given range.

fn iter_natural_lines_for_range(&'a self, range: &Range<CharIndex>) -> LineIterator<'a>