Struct gfx::text::text_run::TextRunSlice

pub struct TextRunSlice<'a> {
    pub glyphs: &'a GlyphStore,
    pub offset: ByteIndex,
    pub range: Range<ByteIndex>,
}

A "slice" of a text run is a series of contiguous glyphs that all belong to the same glyph store. Line breaking strategies yield these.

Fields

glyphs

The glyph store that the glyphs in this slice belong to.

offset

The byte index that this slice begins at, relative to the start of the text run.

range

The range that these glyphs encompass, relative to the start of the glyph store.

Methods

impl<'a> TextRunSlice<'a>

fn text_run_range(&self) -> Range<ByteIndex>

Returns the range that these glyphs encompass, relative to the start of the text run.